Skip idmap1 generation if target defines overlayable
Since idmap1 can't enforce policy restrictions, we can't allow it
to generate an idmap that might get used.

Migrating to idmap2 is too difficult for now as it requires a lot
of PackageManager infrastructure to validate policies, and none of
that is available during manifest parsing, which is where idmap1
is currently used.

Bug: b/130260055

Test: manual test BT issue from b/130260055
Test: manual test app with test overlay toggling Activity enabled

Change-Id: I5e7e321d27aab9be360a5e8be77d804ad3f66336
(cherry picked from commit 6597302f570e37efbef5d86291d41073003ebe5b)