Improved timekeeping logic for USAP Pool refill mechanism.
This patch makes two changes to increase the reliability of the USAP
pool refill mechanism.  The first is to change the storage type of the
result of an expression from int to long.  This will help avoid
potential overflow issues.

The second change mitigates the impact of the non-monotonic system clock
having its value changed.  If the clock goes backwards in time past when
the initial event triggered the refill delay, the delay will now be reset.

Test: Treehugger
Change-Id: Ic6f65de51757803b5ba95975bab3b9322a91c837