Remove VoiceInteractionService whitelisting
Previously, the VoiceInteractionService was started directly
from the ActivityManagerService by temporarily whitelisting its
package.

In this CL, we start it via the VoiceInteractionManagerService.
This takes care to keep the connection to the Assistant alive
(or start it if necessary). This way, whitelisting is no longer needed.

Bug: 119424643

Test: adb shell am force-stop com.google.android.katniss && \
      adb shell input keyevent 219
And check that operation terminates successfully and Assistant is
launched.

Change-Id: I3d44eec5f4f33d33478736c59f0c1ef7a2605979