Clear the EnterTransitionCoordinator only after onResume.
Fixes: 141875256
Fixes: 142224519

In ActivityTransitionState, the mEnterTransitionCoordinator
was being set to null as soon as the transition finished. In
tests, transitions were being finished immediately and the
mEnterTransitionCoordinator was being set to null immediately.
That left time for the same transition to be started again
at a later time.

This waits until the onResume() is called before potentially
clearing the mEnterTransitionCoordinator. This means that all
potential other times where a new EnterTransitionCoordinator
may have been created have already passed and it is safe to
clear the mEnterTransitionCoordinator.

Test: ran ActivityTransition tests
Change-Id: I91978a3c992b533b9074409f05c006fe700d7983