Initialize HDMI PowerStatus to standby
In quiescent boot, the screen stays off until explicit user action wakes
it up. The device should not become an active HDMI source in quiescent
boot.

This commit changes the initial power status of the CEC device.
The CEC device will go into standby PowerStatus initially, and once
boot is completed check if the device is interactive. If the device is
interactive after boot complete, then the CEC device will transition
into the "On" power state as well.

Without this, the following problem occurs:
- The connected TV is using the device as active source
- The device performs a quiescent reboot
- The device reports a power status of "on"
- The TV sends a <Set Stream Path> message to request a streaming path
from the device
- The device sends a wakeUp to PowerManager and becomes Active Source

By reporting that the device is in standby power status, the TV should
not try to request a streaming path from the device.

Bug: 145265927
Test: atest HdmiControlServiceTest
Test: `adb reboot quiescent`, check device reports standby power state
      in `adb shell dumpsys hdmi_control`.
      Check that TV does not request <Set Stream Path> / the device does
      not wake up. Tested with a Sony TV.

Change-Id: I7cae85828b58d482d65f5e645f8940dd92b07f99