Fix opening/closing apps isn't add during device rotating
When animating during device rotation, there may have a timing that
the display is waiting for the orientation change to complete but
the transiting apps won't be added into the opening/closing list by
the waiting state.

It caused when app transition set ready to go, but no opening or closing app can
check if the app has been drawn, check if the app is wallpaper target, or
check if can upgrade to wallpaper transition.

SplitScreenTests#testMinimizeAndUnminimizeThenGoingHome is an example that will hit
this potential problem and failed to update transition from TRANSIT_TASK_TO_FRONT
to TRANSIT_WALLPAPER_OPEN.

The fix is:
- Add DC#okToDisplay(ignoreFrozen) / WC#okToAnimate(ignoreFrozen) methods to ignore
  DisplalyFrozen check in ActivityRecord#setAppVisibility to ensure
  that apps in transition are added to opening/closing list and can
  update wallpaper transition.

Change-Id: I226da85aceca903bd431cc3f90dc2d2c962a632a
Fix: 145151475
Test: SplitScreenTests#testMinimizeAndUnminimizeThenGoingHome