Significantly increase group coalesce time
Manual testing suggests that it can take between 3-20ms for a
notification to make its way all the way through NoMan to us. This means
that groups of a significant size (~25) might be expected to take ~200ms
to completely arrive. Roughly doubling that to account for slower
hardware or a swamped system gets us to ~500ms.

Test: manual
Change-Id: Ia13fc2e88c2c714efa2d46ff98e488627360680c