Refactor SoundTriggerMiddlewareService
This change extracts the client validation as a decorator of
ISoundTriggerMiddlewareService, called
SoundTriggerMiddlewareValidation and leaves
SoundTriggerMiddlewareService to be a mere adapter between an
implementation of ISoundTriggerMiddlewareService and a service stub
and facilities for plumbing as a system service.

This results in a cleaner separation of concerns and also paves the
way for more decorators (logging is planned) to be easily added.

Change-Id: Ia257021d19c5e6ab6509e85a62173a55f2cf8f7a