Revert "Better preload drawable logging"
Bug: 63136392
Bug: 150395371
Test: make

This reverts commit 3bfc729d5c3a5a82677e7ca889c7a7e71ab08593.

Bitmap moving into a new mainline module in S. These module-boundary-
crossing hidden APIs need to be either made official or removed.

We already got the information we need from this logging, so remove it
and the hidden APIs. If/when we need to investigate this in the future,
we should be able to recreate the logging without crossing the boundary.

Change-Id: I5d094a0e427d225719614745867d5b00888d8fec