BLASTSyncEngine: Disable overlapping calls to a single container
Currently we only keep one copy of mWaitingListener and
mBLASTSyncTransaction per level of the hierarchy and so we
are unable to support multiple sync operations at the same level.
For now skip overlapping syncs (making them unsynced). It's relatively
easy for the client to avoid overlapping calls, so this should be
good enough for now. In the future we'd like to keep a map
of mWaitingListener per SyncID, but there is a little
additional complexity about keeping multiple mBLASTSyncTransactions
which we will probably wait for S to tackle.

Bug: 151389300
Test: TaskOrganizerTests
Change-Id: I09efb6bc33468b1e060b8c21bab8cefba65a340e