3/n: Add AuthSession#STATE_AUTH_PAUSED_RESUMING
Updated the "try again" logic to continue using the same AuthSession.
There's no need to create an entirely new one, since all of the
params are the same. When a pausable biometric enteres the
STATE_AUTH_PAUSED state, and the user taps "try again", we enter
the new STATE_AUTH_PAUSED_RESUMING state. New cookies are assigned,
and once all sensors have been returned, the sensors are started and
we go to the STATE_AUTH_STARTED state.

This change also moves most of the remaining AuthSession state
management to within itself.

Bug: 149067920

Test: atest com.android.server.biometrics
Change-Id: Iba80c44faefcf4c4d3f0c51aff635c0a47b9339f