Call proc.setLastActivityLaunchTime() in realStartActivityLocked()
..after we set the time. This is because BAL's grace period stopped
working after that call was removed in ag/9536366 (since without the
method call ActivityRecord.lastLaunchTime would be 0).

I'm not familiar with this area, other than to say this fixes the bug
linked. I'm putting this up given we're blocking some legitimate
activity starts such as starts from the launcher. Please let me know if
this should be a different fix!

Bug: 152621860
Test: atest BackgroundActivityLaunchTest#testBackgroundActivityNotBlockedWithinGracePeriod
Change-Id: I1a329f6a044f5aa491474f5be538991ede0f8f53