Provide a timeout for SurfaceControl
Currently, the timeout for SurfaceControl is set to "-1", which means
that whenever an input event is sent to the view, the view should
respond to the input event "yesterday".

Now that we are actually enforcing these timeouts in ag/10847700, this
is causing any input event going to the SurfaceControl (even the focus
events) to time out.

Set a proper 5 seconds timeout to make these expectations reasonable.

Bug: 143459140
Test: atest android.server.wm.SurfaceControlViewHostTests
Change-Id: I3328b8f71e7f7bd2ccc7a3a64eb25589e68b0f05