Clean up ColorSpace.java
Bug: 156282935
Test: I3ce1bbd4285443a4e481a0a1e9267feafbe49d44
Test: DisplayWhiteBalanceTintControllerTest

Remove cctToIlluminantdXyz, which was @hide, but unused.

Make the @hide constructor package-protected so it can be called by
ParcelableColorSpace, but not from outside the package.

Make cctToXyz and chromaticAdaptation public. These are used by
other parts of the system and may be useful to other clients. Make the
public version of chromaticAdaptation throw IllegalArgumentExceptions
for passing arrays with the wrong number of arguments.

Make mul3x3(float[], float[]) private and copy it over to
DisplayWhiteBalanceTintController. There is nothing ColorSpace-specific
about this code.

Remove Renderer and related. They were originally used to generate the
documentation pages, but they are no longer needed. Remove
ColorSpaceRenderTest, which tested it.

Change-Id: I9094046c8e5e72be0ac208d4d242e6f0b274f386