When entering PIP with multi-activity, re-use mLastNonFullscreenBounds.
In freeform support, we use mLastNonFullScreenBounds to keep track of
the task's bounds in freeform. Such that when maximizing/entering PIP, it
can restore to this bound after un-maximize/exiting PIP.

However when creating a new stack (in the case of multiple-activity in
the task, such as VLC), this information gets lost. We will reuse the
current task's lastNonfullscreenBounds.

Bug: 156780576
Test: In freeform, enter PIP using Multi-activity PIP (e.g. VLC), expand
PIP, the PIP activity is using the same bounds as before

Change-Id: Ic633fd1704c6e909fef0c2bd2dfe4bd9e2ca53ae