Add pinned target indicator to Sharesheet, update visuals
After a user pins an app it takes up a persistent space in the
suggested app row. Users may forget they've pinned an app as
there's no indicator. Add an in-line icon to label suggested
targets as pinned. No indicator should be shown in the A-Z list.
This CL also updates related visual styling to get close to
redlines.

Bug: 150405021
Bug: 152978014
Test: manual
Change-Id: Iee774b88e284617449a673d7e622c0ce233ded1a