Update top fixed rotation launching app when linking new top
Assume the foreground activity is fixed orientation landscape. Recents
(home) is fixed orientation portrait. When swiping the navigation bar
to switch to another fixed orientation portrait activity, before the
recents animation is finished, the fixed rotation app is the recents
activity. And after that, the top activity will be the portrait one.
It will share the existing rotation transform of recents. If the top
fixed rotation app is still the recents, its transform will be cleared
too early (FixedRotationTransitionListener#onFinishRecentsAnimation)
when finishing recents animation that causes flickering.

Also change the usage of mAssociatedTokens to include the owner token,
so that any linked token can clears all rotated state of associated
token at the same time.

Fixes: 157446289
Test: atest DisplayContentTests#testApplyTopFixedRotationTransform

Change-Id: Ibc1c85cf9c9f61297cbcf13412f39b388b874557