Make the bubbles window NOT_FOCUSABLE when the IME is up, but FOCUSABLE otherwise.
This works around an issue where the IME doesn't appear due to bugs with virtual displays + focus. By making Bubbles FOCUSABLE during the initial tap to bring up the IME, we ensure that the Bubbles window has focus and no other windows do. Once the IME is up, we make Bubbles NOT_FOCUSABLE so that swipes on the Bubbles UI doesn't steal focus from the ActivityView.

Also, adds a touch listener to the expanded view that prevents the stack from collapsing if the side padding is touched during a back gesture.

Test: manual, lots of IME mashing
Bug: 156785479
Change-Id: Ia9de170a9ce4c9b99017d0ce6901621f70b414fc