Fix NullPointerException on estimate object
The severe low battery warning notification isn't working if a device
which uses time based estimates doesn't connect to Wi-Fi or mobile data.
This is caused by a null estimate object which is obtained from
EnhancedEstimates. This change is to make EnhancedEstimates return an
unknown estimate and add checking time remaining condition to see if
time remaining is valid. AOSP behavior is unchanged and this only
affects devices which use time based estimates.

Bug: 152880156
Test: atest SystemUITests
Change-Id: I8f8ac67ebe66ffd748d958d34f9b31324c52f681