Fix manage not being dismissed by taps in the AV bounds.
This actually isn't a touchable region issue - the expanded view steals touches from the stack if they're within the AV's vertical bounds so that back swipes across the side padding aren't intercepted by the stack, causing it to collapse.

We should be checking the x coordinate as well, since the stack *should* be able to intercept touches from the AV itself.

Fixes: 158674822
Test: manual
Change-Id: Ieb12e6d97a0c180139bcaa1634e0ce47899d7c72