Allow doze dream package to start a DreamActivity
After we made the DreamService use an activity, we introduced checks
that only the current active dream component can start a dream activity.
This meant that if the doze dream is in a different package, it would
get suspended.

This CL fixes that by whitelisting also the doze dream package.

Bug: 160109097
Test: m && flash && put two different dream as doze and active dream &&
      check that the doze starts
Test: atest DreamManagerServiceTests
Change-Id: I553dd1d44f3303d18278482b4a4e88a5564a94d3