Use vibration id as callback reference in VibratorService
Instead of creating a new global reference for each Vibration jobject
for native callback we are now only keeping a local copy of a jlong
vibrationId to be passed back to a method from VibrationService.

The single global reference to VibrationService is created in
vibratorInit and used in all callbacks.

Fix: 163901838
Fix: 165045639
Fix: 164290392
Bug: 153418251
Test: atest FrameworksServicesTests:VibratorServiceTest
      atest CorePerfTests:ViewPerfTest
      atest CorePerfTests:VibratorPerfTest
Change-Id: I91791d0a8c524f09167bf6256faa751b957f3056