audio service: add watchdog for audio mode
Add a watchdog to remove an application from the audio mode
owner stack if it has requested mode IN_COMMUNICATION and
has not been playing or capturing audio for more than 3 seconds.

This will prevent volume, routing and recording issues when
misbehaving apps do not properly reset the audio mode at the
end of a call.

Bug: 161564483
Test: Manual regression tests for VoIP use cases
Change-Id: I7c6f57a2c9a6c31d107a1bba2d7dfac6e3f275bd