Fix some blast issues with SurfaceViews
1. When calling reparentChildren, we want to move children from old view
root surface to new view root surface. If blast was enabled, the current
code would mvoe children from old view root surface to the layer above
the new view root surface. This would cause the SurfaceView to be
parented to the wrong layer

2. When the view root surface is changed, VRI is supposed to notify the
SurfaceViews so they can call setRelativeLayer to the new view root
surface. This wasn't getting called when blast was enabled because the
logic to check if it was a new surface is different when using a blast
layer. In addition the checking the sequence id that we do now, also
check the relayout flag RELAYOUT_RES_SURFACE_CHANGED since that should
indicate that the surface was changed.

Test: Split screen with Youtube.
Change-Id: I1012678b60cec87a702e4d0ffe3e197eda66cff6