Fix users' permissions with bg mic/camera
When the background mic/camera changes were submitted the system started
exempting preinstalled apps to use the permissions. The state can be
corrected since the erroneous exemptions have been fixed.

This can take a couple seconds so it's been decided to fix the state
after boot, this may cause an app to be killed while using it but this
would be very rare and there's only one chance for it to happen.

This change should be reverted after an update is pushed. Fairly
verbose logging (which should only happen at once on boot) is being
added so any issues this causes can be easily debugged to avoid leaving
users in a bad state.

Bug: 169856898
Bug: 158311343
Test: Push build without this change, delete permissions db, boot,
          assert bad permission state, push this change, boot
Change-Id: I7f2729b3f5ce19759548969f54632f8b08be8261