[pm/incremental] disable unstartable state and hide related APIs
Per discussions in go/incremental-when-to-freeze, we will not mark an
app unstartable, regarless of streaming health metrics.

This CL is a temporary disablement. The startable state will remain true,
and there will be no broadcasts about startability.

The event tracking and health monitoring is kept as is.

If we get feedback and need to put it back, it'll be a small change.
Otherwise, we will clean up the related code.

+ also fix storage health monitoring params

BUG: 171920377
Test: builds
Change-Id: Idf7d32d0ca3a6e08c9673914f27f22b531b91ac2