Revert "Enable building libandroid_runtime for Windows"
Building libandroid_runtime for Windows means building the
libnativehelper dependency for Windows as well. Since that library is
moving into the ART Mainline module it would mean having to provide it
as a prebuilt.

By removing this we avoid that, and the Windows variant isn't required
on the main branches, only in a downstream one.

This reverts commit e07814656a82b8def7d0d98f8a067fa633eabd2c.

Test: `m` on git_master
Bug: 172649321
Change-Id: I5f0162e9b87529ee8b339bd24e5dee3e9bdd23a8