Change shouldShowIme -> displayImePolicy
This CL allows for a Trusted Display to specify not only whether the IME
should show on itself or the fallback Display, but also whether the IME
should not show at all.

This behavior is useful if interacting with a display that has a built
in IME; it will prevent an IME from also showing on the default Display.

Bug: 170233231
Test: atest MultiDisplaySystemDecorationTests#testDisplayPolicyImeHideImeOperation
Change-Id: Id86db922e8f2c3ca065f830ee371cccaa7c36101