Implement uid logging in the messageReported atom
This requires storing the calling UID as the work source UID
to preserve it on Runnables executed on the service thread.
This was done by refactoring all methods like
HdmiControlService#runOnServiceThread to preserve the work source UID
during execution of a Runnable.

Test: atest com.android.server.hdmi
manual using ./out/host/linux-x86/bin/statsd_testdrive

Bug: 176803456
Change-Id: I06718e8a9104c40e3b895427677c4bcf0945ccf2