Ensure size-compat check is correct in LaunchParamsModifier
Currently even if "Freeform SizeCompat Mode" is enabled, size-compat
apps are launched in maximized. This is because
shouldUseSizeCompatMode() returns false in TaskLaunchParamsModifier.

The reason for this is the activity type is still undefined at this
point as if the activity should be standard, it's set when it's
added to its parent task (See Task#addChild). For other non-standard
types, it's set immediately in the constructor of ActivityRecord.

Thus it should be safe to consider ACTIVITY_TYPE_UNDEFINED to be
size-compat, in which way, we can properly handle size-compat apps
when resolving the launch params.

Bug: 177630091
Test: go/wm-smoke
Test: Size-compat apps are launched in freeform with the dev option on.
Change-Id: Ic6cf1b4fbeee20483d819c22fba7769966931145