Run alignment logic of menu icons when the menu is created
The algorithm's icon arrangement logic was previously ran in
onBoundsChange, which could cause the icons to be misaligned
the very first time the menu is shown after being created.
This change ensures that the algorithm does an initial run
when the menu is first created.

Bug: 178708863
Test: manual - open YT PIP, tap to see menu, see the gear
icon is aligned to the top-start corner

Change-Id: I2cb53578db897af62be7ebd3ed40946e16606568