Include knownCerts in Parcels for PermissionInfo and ParsedPermission
Android 12 adds support for a new knownSigner permission protection
flag that allows one or more trusted signing certificates to be
specified with the permission declaration; a requesting app signed
by any of these certificates will be granted the permission. This
commits adds the new knownCerts Set to the Parcels for
PermissionInfo and ParsedPermission instances.

Bug: 172970984
Test: atest PkgInstallSignatureVerificationTest
Test: atest PermissionInfoTest
Change-Id: I07e415c0d3df6fe352ec5b04a29d986a2647fa19