Reduce unnecessary config propagation of default display
Global config is also the override config of default display.
When updating the override config, the hierarchy is already notified.
So it doesn't need to update twice. Previously, the first update
from mRootWindowContainer.onConfigurationChanged only recomputes
the old config again because the override doesn't change.

This may reduce up to 3ms (depends on the amount of window
container in default display).

Also remove unused parameter deferResume.

Bug: 159103089
Bug: 178472794
Test: CtsWindowManagerDeviceTestCases
Change-Id: I1437a26e435666c13faaf27e0b0ef81adf739fe0