Make "adb shell cmd input_method tracing" official
This is a follow up CL to our previous CL [1], which introduced a new
shell command to start/stop IME tracing.

Perhaps this might be too opinionated, but "adb shell ime" had been
used as a convenient wrapper only for enabling/selecting IMEs and not
for doing anything beyond that.

Admittedly

  adb shell ime tracing start

is short and convenient, but keep in mind that for winscope we already
type the following command.

  adb shell cmd window tracing start

For better consistency, it'd make mose sense for us to use the
following command instead.

  adb shell cmd input_method tracing start

Note that this CL does not remove "adb shell ime tracing start" so as
not to break internal tooling, but at some point we should migrate all
the tools to use the new command.

 [1]: Ia89f11d5ef8a220ea7746191b18769cea5a8359d
      d96fc1dce576000de6d535fc67cd8cb5cbcad281

Fix: 180765389
Test: "adb shell cmd input_method tracing start" works.
Test: "adb shell ime tracing start" still works.
Test: "adb shell ime list" still works.
Change-Id: Ic53fd7c8baa3340250cf5e03bd225cce219eae61