Fix revokePermissionsNoLongerImplicitLocked() for shared UIDs.
The method should have considered all the packages in a shared UID, but
it only took the package passed in into consideration, and would
incorrectly revoke permissions that's implicit from another package in
the same UID. The proper way is to always consider permission state by
UID, but as a simple fix, considering each package only by their
requested permissions works the same. The revocation of permissions no
longer requested is handled elsewhere and involves removing the entire
permission state, so it's also safe to only consider requested
permissions here.

Bug: 183304624
Test: manual
Change-Id: I26b4bd7007cfed662c56628723e4be2718093880