Call surfaceplacement if the window goes for relayout
We do not perform surface placement on the window if its considered
gone for relayout. This change was introduced to prevent us from
using invalid mFrames. However there is a discrepancy in the gone
for relayout check. We will still relayout a window if the window is
not currently visible but a layout is request of reasons such as
window size changed. Fix this by requesting a surface placement
if relayout is requested.

This fixes an issue where a toast window created under the lock screen,
shows up with the wrong position for a few frames.

Test: go/wm-smoke
Test: open toast under lock screen, check toast appears in the correct position
Fixes: 175668342

Change-Id: Ibfe6aeb5704a348014e2ebca996874ebae18fdd2