Postponing package rescan to prevent blocking device startup
ShortcutManager currently enforces a rescan of manifest shortcuts across
all of the packages after OTA update. This would lead to significant
delay before launcher is visible to the user when user tries to unlock
for the very first time after the OTA update.

This CL moves package rescan to a later timing (i.e. when non-pinned
shotrcut is requested) to mitigate the delay.

Bug: 151359749
Test: atest ShortcutManagerTest1 ShortcutManagerTest2
    ShortcutManagerTest3 ShortcutManagerTest4 ShortcutManagerTest5
    ShortcutManagerTest6 ShortcutManagerTest7 ShortcutManagerTest8
    ShortcutManagerTest9 ShortcutManagerTest10 ShortcutManagerTest11
    ShortcutManagerTest12
Test: atest CtsShortcutManagerTestCases
Change-Id: I5f4757e660a6e971974f3ca530a81fb25d21828c