Re-order early return when targetPkg is null
It's possible that the target package is on removable stroage or
uninstalled with keep data. The caller should be able to get the
related info of the target with flag MATCH_UNINSTALLED_PACKAGES.
However, the caller wouldn't be able to get the valid info if we
directly filter out when targetPkg is null early.

Fix: 187650819
Test: atest AppEnumerationTests PackageVisibilityTest
Change-Id: I7c6b2dcdb77c543e0826e5dff0ddc9f66e489033