Avoid potential deadlock
There has been a few reports indicates ANR was observed  when calling
ShortcutBitmapSaver#getBitmapPathMayWaitLocked, which timed after 30
seconds.

This CL reduces the timeout to 5 seconds as 30 seconds is unreasonably
long. Additionally, since ShortcutService#postValue is no longer needed
for AppSearch migration, it is better to remove this api to avoid a
potential risk of deadlock which might have caused the ANR in the
first place.

Bug: 189840177, 189861955
Test: atest ShortcutManagerTest1 ShortcutManagerTest2 ShortcutManagerTest3 ShortcutManagerTest4 ShortcutManagerTest5 ShortcutManagerTest6 ShortcutManagerTest7 ShortcutManagerTest8 ShortcutManagerTest9 ShortcutManagerTest10 ShortcutManagerTest11 ShortcutManagerTest12
Test: atest ShortcutManagerClientApiTest

Change-Id: I33fe05a08df43e926904de4420f863f65d3cbad8