Fix Bouncer-Showing Status.
Prior this change, if you drag up slowly on the lock screen,
(no password/pin), it will mark the bouncer as showing because
it thought the bouncher "would" show. The Bouncer never actually shows,
however, so we get stuck in that state.

With this change, we ensure that the bouncer is never marked as showing
in that scenario.

This fixes an unexpected visual artifact where the sim-pin error messsage
would show on top of HUNs because it was left visible when the Bouncer
was marked as showing.

Fixes: 191733743
Test: manual
Change-Id: Ia2e7dc72ca57ae932cb58e99410daffc9e7c0672