Move temporary failure logic up the hierarchy
BinderLocationTimeZoneProvider can temporarily fail if the remote object
dies. This was handled in BinderLocationTimeZoneProvider, but it is
being moved to LocationTimeZoneProvider in this commit.

This makes it easier to fix a bug (also fixed in this commit) where the
initialization timeout was not cancelled if the remote peer died during
initialization. This would later result in a logged warning when the
timeout fired and the LocationTimeZoneProvider was no longer in the
"initializing" state. This has only been witnessed during unit tests
while debugging.

Test: treehugger only
Bug: 188780558
Change-Id: I76a22edaf9bea4287d07c408ad43fc3e5a8e49fc