Dismiss SIM PIN screen when eSIM disabled
eSIMs provide a mechanism for disabling directly from the SIM PIN
unlock screen. When tapped, keyguard receives an event with a
subscription id of -1, which an invalid subscription id and is mostly
ignored. However, in the case where there was a previously valid id,
keyguard should issue a callback to notify consumers of this state.

Fixes: 191432304
Test: atest KeyguardUpdateMonitorTest
Change-Id: I455d7a1fcde6cbfbf6914f4900eb58cf39bfd715