Make sure SIM PIN screen shows
It is possible to receive SIM events that alert the keyguard to show
the SIM PIN unlock screens. However, if these events are received
after keyguard has been told to dismiss but before it is actually
dismissed, this event gets ignored.

The fix: add a flag to look for this state. If we get the keyguard
dismissed event and the bouncer has not shown, bring back to the
keyguard.

Fixes: 192048149
Test: atest KeyguardViewMediatorTest
Change-Id: I43de785252827d78a86d075ee7ec0861778e3360