Accessibility for privacy status animations
This CL adds contentDescriptions to the privacy event chip and the
ongoing privacy dot views. This is a partial fix which allows talkback
for the ongoing privacy chip, and allows talkback scrolling to the
privacy dot. However, due to the fact that the dot lives in a different
window, the privacy dot shows up in the wrong part of the accesibility
tree and still needs to be virtually placed in the status bar.

Bug: 187197696
Test: manual
Change-Id: I6210e18c75a1aab61cbcf619a9a31c459fffe544
(cherry picked from commit 9b94acf5b95a1a5a451d49f42c7e2a33fe5790ba)