Preserve invocation order in RemoteInputConnection
This is a follow up CL to our previous CL [1], which introduced
early-exit paths to some RemoteInputConnection methods to protect
innocent IME clients from unexpected process crashes when an IME is
calling InputConnection APIs with invalid parameters.

Although protecting IME clients from crashes still makes much sense,
implementing it as an early-exit style in RemoteInputConnection may
expose observable inconsistency to IME developers in terms of the fact
that InputConnection#getText{Before,After}Cursor() can also work as a
fence operation that would not return until all the previously issued
InputConnection API invocations are handled in the IME client side.

With this CL, the following methods start behaving as a fence
operation even when an invalid parameter is passed, by checking the
parameters in the IME client side.

 * RemoteInputConnection#getTextAfterCursor()
 * RemoteInputConnection#getTextBeforeCursor()
 * RemoteInputConnection#getSurroundingText()

There should be no performance impact for IMEs that do not make such
an invalid (and unnecessary) API calls.

 [1]: I95169735198f8363c981a61e20234dfebfd645b1
      1dfa38c929dae55c1753c6e7a97b840b0d108a16

Bug: 169114026
Fix: 194110780
Test: atest CtsInputMethodTestCases:InputConnectionEndToEndTest
Change-Id: Ie0c18d0c9b8bf8f02f2fcdca5aac7e580c6bf2cd