Move PackageUserState into system server
This moves the main interface, which allows all internal
and eventually mainline APIs to live in the same place and
reference each other in code and docs.

A compatiblity interface called FrameworkPackageUserState was
maintained in the old location because a lot of _Info generation
code relies on it to wrap simple booleans and other getters.

Refactoring those methods to use direct parameters rather than
the interfaces was deemed not worth the effort, as there would be
a large number of new parameters and method calls to fix up.

This can be improved over time and maybe once parsing/_Info object
generation is removed from framework.jar, this compatibility
interface can be fully removed.

Bug: 201811826

Test: atest PackageUserStateTest

Change-Id: I23ab9be440f2bcf5bcf58fea42a355caf2c13933