Attempt 2 for adding OWNERS for SNTP code
Commit 220acd7cd884e15f5b671cd13adab14614cf4cb7 combined with pending
reviews demonstrated the OWNERS syntax has quirks and it is not working
as intended.

Documentation used:
https://android-review.googlesource.com/plugins/code-owners/Documentation/backend-find-owners.html#syntax

The syntax guide recommends that patterns like sub-dir/* aren't used as
they can cause problems when renaming directories; better to have OWNERS
in the sub-dir. The pattern sntp/* didn't appear to be working (perhaps
because of spaces).

Test: None
Bug: 199481251
Change-Id: I193744eaaad529b9680d3ccf905b86d68f62f7ff