Refresh CompanionDeviceManager's API surface
Make the AssociationInfo and DeviceId classes public.

Deprecate the existing public associate() method that take a callback
based on onDeviceFound(), and introduce a new one that takes an
AssociationRequestCallback.

Deprecate the existing public disassociate() method that takes a MAC
address, and introduce another one that takes an associationId.

Deprecate the existing public getAssociations() method that returns a
list of MAC addresses, and introduce a new public method
getMyAssociations() that retuns a list of AssociationInfo objects.

Introduce new system API to retrieve list of all associations for a
user: getAllAssociations().

Introduce new system APIs for monitoring changes to the existing
associations: OnAssociationsChangedListener interface, and
add(remove)OnAssociationsChangedListener() methods.

Bug: 194301022
Test: atest CompanionDeviceManagerTest
Test: atest BluetoothInstrumentationTests
Test: atest SettingsRoboTests:NotificationBackendTest
Change-Id: Ic0235db29b26ed1a427c71edb2eb0df76fa4dc4c