Fix UDFPS not working when a HUN is showing
We didn't want the UDFPS overlay to handle touches when a user pulls
down the notification shade. We accomplished it by ignoring UDFPS
touches if the notification shade is expanded.

However, that had an undesired side effect. Apparently, the notification
shade is considered expanded when a heads-up notification (HUN) is
showing. Which means if a user receives a HUN during authentication,
their touch will be ignored.

This CL fixes the undesired behavior by also checking the expansion
ratio. It appears that when a HUN is showing, the notification shade
reports itself as expanded with the 0.0f expansion ratio.

Bug: 203213357
Test: authenticate in BiometricPrompt with a HUN showing
Change-Id: I597393b8f8bafda3b28e54237a44ba7517681869