enc stat: Add feauture and its keys to MediaCodec
Add feature name 'EncodingStatistics' to MediaCodecInfo and its two
keys of average_qp and picture_type to MediaFormat.

App client chooses the 'Level' of an Encoding Statistics information,
where the Level decides the scope of encoding statistics that video
encoder provide the App client with.

With Android T, there will be one level only and the expected kinds
of encoding statistics are average block QP and the picture type of
a frame.

Bug: 205793764
Bug: 203696003
Test: cts
Change-Id: Id8c61de9646ac03926f00a9d223b022bbb9a6925