Do Not Report Activities that are Finishing
...in TaskFragment

There may be a delay between when an activity is marked
to finish and when the WM hierarchy actually removes the
activity from the TaskFragment. In between this delay,
the TaskFragmentInfo will change because the runningActivityCount
will change, triggering an onTaskFragmentInfoChanged.
This callback is unncessary because there will be another
callback soon once the activity is removed from the hierarchy.

Bug: b/205901017
Test: atest CtsWindowManagerJetpackTestCases:ActivityEmbeddingLaunchTests
Change-Id: I64e08acc2a745f6f242996e99240fe4a8f0dea97
Merged-In: Id773ac3f0afea198ad4f8b4594d7a4cf0ad98ea8