Prevent double keyguardGoingAway, and finish cancelled animations.
keyguardGoingAway could be called twice in a row due to race conditions between tryKeyguardDone and hideLocked.
This resulted in the first animation being cancelled. This would normally not be a problem, except that we also
erroneously did not call onAnimationFinished() on the remote animation callback if the animation was cancelled,
which meant WM waited the full 10 seconds before timing out and showing the app/launcher.

Fixes: 208731860
Test: atest SystemUITests
Test: repeatedly UDFPS and wait for a cancelled animation, see that it becomes visible
Change-Id: I6d3401e85df99a2a9be75cb51fca7f7c5ee939d4