Remove check preventing keyguardGoingAwayRunnable from being called if it's already going away.
This check was added in ag/16441222. The root cause of ag/16441222's bug was that we were not
handling remote animation cancellations properly, so it fixed that and also added this check to
prevent extraneous cancellations from happening in the first place.

That check was not necessary to fix the original bug (since we handle cancellations properly now,
including extraneous ones), and appears to be causing issues with camera CTS tests and at least
one report of a stuck lock screen. I can't reproduce any of that locally, but given that it was
not necessary, we should remove it pending investigation.

Bug: 210831136
Bug: 210732858
Test: atest SystemUITests
Change-Id: I78af499016e9c46f881fb7045e9823273162ddcc
(cherry picked from commit 44add4590abd06f7e8816c738bf8f9b5c04e4945)
Merged-In:I78af499016e9c46f881fb7045e9823273162ddcc