TIAF: add permission BIND_TV_INTERACTIVE_APP
Similar to BIND_TV_INPUT, in the manifest declaration, specifying
the BIND_TV_INTERACTIVE_APP permission to allow the service to
connect the TV interactive app to the system.
A system service performs the binding and has the
BIND_TV_INTERACTIVE_APP permission.

Link of BIND_TV_INPUT as a reference:
https://developer.android.com/training/tv/tif/tvinput#manifest

Bug: 215773542
Change-Id: Id185e0495420e67b2566a848dd81e0f72ad6d083