New pipeline: defer visual updates when collapse
In the old pipeline, StatusBarNotificationPresenter had logic which
deferred changes to the notification shade if the panel was collapsing,
or if an activity was launching.

This CL introduces a change to the new pipeline to replicate the old
behavior. A new API is added to the NotificationStabilityManager,
isPipelineRunAllowed(), that is checked before performing any visual
updates to the notification shade.

isPipelineRunAllowed() is driven by listeners to collapse + activity
launch events emitted from the NotificationPanelViewController. When the
value of isPipelineRunAllowed() would change, the
NotificationStabilityManager is invalidated.

Test: launch activity from notification, verify notification isn't
      removed while activity is launching
Test: atest VisualStabilityCoordinatorTest
Fixes: 205994866

Change-Id: If9921353ee7393a978a84f016ba62856fa13afc4