Remove InputMethodManagerService#mCaller
Having both HandlerCaller and Handler in InputMethodManagerService has
confused us especially about which one should be used to obtain a
Message object.  This is really confusing when a Message can be
removed with .removeMessages() because it behaves like each Handler
has its own namespace.

Let's unify our usages into Handler in favor of simplicity.

This CL changes nothing in terms of how InputMethodManagerService
behaves unless there was any misuse of removeMessages() and we had
relied on such a bug.

Bug: 192412909
Test: presubmit
Change-Id: I63a3d0015ddc937f5f6928cebcad893c75e7863c