Reset NotificationPanelView when pressing power button before fold to AOD animation
When we hit power button during folding: before
fold to AOD transition starts but after we prepared
the starting params of NotificationPanelView it remained
in this state (alpha = 0, translation = -x). So it led
to an issue that user don't see the keyguard.

Added logic that resets this view to its original
state when we are waking up and the animation hasn't
started yet.

Bug: 215703534
Test: fold and quickly press power button
 => check that keyguard is visible
Test: fold to AOD and press power button => keyguard visible
Test: lock/unlock using power button
Change-Id: I88b79aca462137723e174f2f4196cbd08d71f664