Lock default display when it sleeps independent of global wakefulness
When dealing with multiple PowerGroups, we calculate the global
wakefulness as the highest of any individual group. In the case where
a non-default PowerGroup (e.g. for a VirtualDisplay) is awake and the
default PowerGroup transitions from awake to asleep, we need to let
the PhoneWindowManager know so it can tell Keyguard to lock the
default display.

Fixes: 208229988
Test: atest PowerManagerServiceTest
Change-Id: Ib0fc4fdc30ca6f111e3c9d428a45bab5d403ad15