wifi: Fix unit test error
The default timeout value is changed to -1 after T.
Since the migration module is not a part of mainline.
So fix unit test it directly for T unit test error.

For old device, it should still use old unit test version with old value and get pass
because the getShutdownTimeoutMillis() returns old value on old build.

Bug: 213289672
Test: atest -c android.net.wifi.SoftApConfToXmlMigrationUtilTest#testConversion
Change-Id: Ia14b019101b8e78fc3f7374c907453508dfea420