remove redundant share intent for Wear bug report
In ag/13068228, we amended the bug report functionality for Wear to
include a share intent (ACTION_SEND_MULTIPLE) which would be sent once a
bug report is complete. However, this intent is not required as Wear
sends the bug report directly to WCS via TriggerBugReportService in
ClockworkShell.

TriggerBugReportService listens for the bug report finished action
"com.android.internal.intent.action.BUGREPORT_FINISHED" and sends the
bugreport files to WCS. A chooser activity is not launched on Wear.

This issue surfaced due to a silent FileNotFoundException in the logs.
This occurs because the bug report files are consumed and deleted by
TriggerBugReportService in the standard Wear bug repot flow.
By the time this share intent is read, the files no longer exist,
leading to the exception.

Most of the flow is illustrated in the following screenshot:
https://screenshot.googleplex.com/8kjMkirzvPyFHg7

The changes here does two things:
- remove code that introduced the share intent
- remove redundant code after above change
- add missing break statement

BUG: 208822859
Change-Id: I38d1ee961caf2eead105acb9e84101eef45df9b0