Fix the new QS footer buttons animation
This CL fixes the animation triggered when opening an app/dialog from
one of the new QS buttons.

Before this CL, the background of those buttons was oval and therefore
can't animate its radius.

This CL changes the oval shape by a rounded rectangle so that it can
properly animate the radius of the background.

Bug: 217196619
Test: Manual
Change-Id: I77ebbb9be62b3457844e06803ab3ce7b9af5d594