Fix app visibility gaps from AppOps
The app visibility enforcement in AppOpsService throws IllegalArgument-
Exception if the incoming package is installed, but not visible to the
calling UID. This enforcement wouldn't take effect when the package is
not installed. And then, the following logic would return MODE_ERRORED
instead. The difference gives malwares the chance to do a side channel
attack, and it leaks installed packages.

This CL treats invisible package the same as not installed, and return
the same result instead of throwing the exception. It also handles the
special callers since their packages don't actually exist.

Test: CtsAppOpsTestCases
Test: FrameworksMockingServicesTests:AppOpsServiceTest
Bug: 214894893
Change-Id: I9453bea1835f4dc45010ba72f00b9ad017e5788e