Tie game session lifetime to game process lifetime.
This change adds a process observer which watches for game process
activity and death. Process activity is tracked so it is known
which process is associated with which packages. When a process dies,
if it is the only process for a game with a game session, then the
game session is destroyed. If a process for that game is later seen
again, then a new game session is created.

Bug: 204504136
Test: Manual tests

Change-Id: Ic2a264c85867c17ab84de855240f2bba3e70a89d