[Media TTT] Refactor media states to use an enum instead of sealed classes.
There ended up being a lot of different constants associated with each
state and a lot of different maps or `when` functions in different files
to associate the constants with the states. It's more straightforward
to define an enum for each class. (And this will be even more true
when we add constants for the UiEvents.)

Bug: 203800663
Test: manual: verify adb commands still display all the different chip
states
Test: media.taptotransfer tests

Change-Id: I01d0ba2654d1f78fca1f697f81d0d6fa98aa2513