Rework system event animations to a federated model
Before this cl, all system event chip animations were structured to run
via callbacks from a single ValueAnimator which ran from [0,1], and
clients could implement either the system chrome callbacks or the chip
callbacks.

This change reworks the whole system to instead use the animation
scheduler to collect animators from any relevant clients and play them
all together. It allows for much more granular control over individual
elements that want to participate in the animation.

The tradeoff here is that there is no single place for the animation to
be coordinated from. The entire experince needs to be composed by
orchestrating every animation individually and testing manually to
ensure that the motion is adequate.

Test: manual
Bug: 187881849
Change-Id: I8b2ec28b0489f6ec5870e4e14e5e9fd1bd15e076