SystemUI: remove unused code that showed radio status in CryptKeeper mode
Since support for Full Disk Encryption was removed in favor of
File Based Encryption, "CryptKeeper mode" no longer exists and the
"vold.decrypt" system property is never set.  Therefore, remove the
unused SystemUI code that showed the radio status in CryptKeeper mode.

This is, very roughly, a revert of ag/1350315 which added this feature
in 2016.  However, the code has changed a bit since then.

Bug: 208476087
Change-Id: I6231762e8353113366eba92b686d6d61582fb2b7