Consider requested orientation for snapshot compatible
For example, the display is portrait so the task is also portrait.
But the activity in the task requested landscape which will make
the display update to landscape later. Then if the snapshot was
taken in landscape, the snapshot should be compatible because
the activity will be pre-rotated with fixed rotation.

Bug: 216079545
Test: ActivityRecordTests#testIsSnapshotCompatibleTaskSizeChanged
Test: Swipe quickstep to an existing activity with different
      orientation. There should not be splash screen window.
Change-Id: Ic87ead99c918de19438f88472f01541208469a36