Updating QS state to default when going from or to split shade
When going from split shade to regular shade Quick Settings now will be collapsed. When changing to split shade, they will be expanded - which was already the case with UI but internal state of NPVC was not reflecting thatand it caused most bottom tiles to not work after rotating to landscape. Changing to default state will likely prevent many nasty bugs in the future.

Also renaming `mShouldUseSplitNotificationShade` to `mSplitShadeEnabled` which is nicer, shorter and better follows the usual convention of *Enabled.

Bug: 228198572
Fixes: 230612123
Test: 1. Rotate from split shade to regular shade and see QS are collapsed
2. Rotate from portrait with collapsed QS to split shade and see most bottom tiles are working

Change-Id: I13eedf62b92dd5eb06056739f3950c76b6601a07