Remove the ActivityLaunchAnimCoordinator.
This coordinator was recently added for the purpose of extending the lifetime of the notification entry while the launch animation played to prevent an entry instance mismatch crash when the app cancelled and reposted it on app launch.  Unfortunately, this wasn't a scalable solution: 1) there was another launch scenario that wasn't covered by this animation, and 2) if the notification was auto-cancelled and reposted, this would not work because lifetime extenders are not allowed to extend the lifetime of user-dismissed notifications.

That bug and these new crash cases were solved by alerting the NotifCollection when the animation starts, and then not dismissing the notification when it ends if it was already dismissed by the system server. That is implemented in Ia66ae5ade3fbfdd0436d54dcbeef720618622716, but makes this coordinator and the events it listened to unnecessary.

Bug: 230540148
Bug: 227254780
Test: manually validate that crashes are still fixed
Change-Id: I28f7226a30df7730dbae264bfd101397293a9a03