Update VPN app exclusion list when packages are added or removed
This commit passes through package added/removed broadcasts,
triggering the VPN to update the allowed/excluded UIDs.

This is required because the VPN app exclusion list is stored
in the keystore. The UID range for the VPN network agent should
be updated when the package status is updated.

As an example, consider the following sequence:
- Package A is installed, and in the exclusion list.
- A is uninstalled
- VPN starts, A is not installed, so no UID A to exclude
- A is re-installed after VPN is started
- The VPN would include the app A, potentially breaking it.

Bug: 231373589
Test: atest FrameworksNetTests
Change-Id: I7ee32103ffdfc97e4c95007f35671cefeb9b22d7