Log the attachment index
We log the index when moving a view but not when attaching.  This is an incredibly low-risk way to have better indications of when unexpected views (such as the silent header) exist in the shade.  It will also add information about where views were attached without having to enable verbose logging.

Bug: 224929774
Bug: 234824085
Test: dumpsysui NotifLog
Change-Id: Iac7a9d4e6fe505050a1a55afe78a8ead46af37a7