Fix pinched short notification at bottom of lockscreen.
This change disables the feature on the lockscreen which smoothly transitions the roundness of notifications when they get closer to the shelf.  For short notifications (a common occurrence when notifications are redacted) this would cause the TOP corners of the bottommost notification (and the bottom corners of the one above it) to be rounded on the lockscreen.  This was because the 'shelf' start was considered to be a shelf height ABOVE the bottom of the stack.  This CL just tweaks the shelf position down to where it intuitively is.

Fixes: 235234674
Test: manual
Change-Id: I2da67022a62f0b3adcd00df1c9b372754ee643fc