Attemp to fix Launcher stay invisible after keyguard unlock.
For the keyguard exit animation end up in a fully/semi transparent
case, because the timing of keyguard exit animation may not match to
panel fling status, so no matter it's playing fadeInSurfaceBehind or
surfaceBehindEntry, there should always cancel the animation when
received notifyFinishedKeyguardExitAnimation.
Another case is that the Notification shade become invisible while
animating, in this case the SyncRtSurfaceTransactionApplier won't apply
next transaction because there won't draw any frame for that view.

For Launcher re-show case, because the transaction cannot be applied
synchronized at KeyguardViewMediator#handleStartKeyguardExitAnimation,
so we should initialize the alpha to zero for the opening app.

Test: enable shell transition, then scroll up to unlock device,
a. verify the launcher won't stay in a fully/semi transparent.
b. verify the launcher won't reshow when scrolling on keyguard.
Also verify above cases via fingerprint unlock.
Also verify when unlock to split screen.
Also verify nothing change for legacy transition.
Bug: 232852578

Change-Id: I520102f476a417b66822d44f8effb905da4ff943