Set mAlreadyRequestedSync only when in a BLASTSync
WindowState#useBLASTSync returns true if in a blast sync but also if
there are mDrawHandlers. In that case, we only want to mark
mAlreadyRequestedSync to true if we know it's a blast sync since the
unmark is done when the sync is complete.

Test: Pip expand no delay
Fixes: 236882306
Change-Id: Ifb5785b93888a0fb3e27180005ea1b3940a293b7
(cherry picked from commit 9aed73773a8e754a0bf9dc17ffc55169ae2fdda8)
Merged-In: Ifb5785b93888a0fb3e27180005ea1b3940a293b7