Mark display ready when it loses active child
When reparenting an actively-collected container out from one
display to another, when it becomes ready it will notify the
destination display and not the source display. However, since
it was added to syncset on the source display, that display
will have a ready-tracker entry (since we currently track
readiness per-display -- like legacy transitions). As a result
the transition will time-out since the source display will
never be set ready.

To fix this, mark the display ready when a child is moved
to the other display. The transition, in total, will still
remain unready until that child is ready (on the new display)
so it won't fire early.

Bug: 237545923
Test: atest MultiDisplayPolicyTests#testStackFocusSwitchOnStackEmptied
Change-Id: Ida70bee427e1baa517fa231ae52c53859c2f3dba