Change the strategy of handling batched input events
This CL changed DragResizeInputListener to use the same strategy of
handling batched input events as ViewRootImpl. This effectively
throttles the resize requests by the refresh rate of the primary
display.

This CL also stops relying on WM core responding to WM shell on each
size request, which may not be always true depending on the WM core
state, e.g. when WM core detects the requested bounds violate the
smallest dimension requirements.

Bug: 239238690
Test: Fluid resizes work with both fingers and mice.
Change-Id: Iad95ba7a5163138c52d902c94e1a3e4dc3180b6b