ArrayUtils: rename concatElements() to concat()
concatElements() was probably named as it is to be like addElement() and
removeElement().  However, that doesn't really make sense because the
parameters of concatElements() are only arrays, not array elements.

Rename it to concat(), as that is just as clear and follows the example
of Guava's ObjectArrays.concat().  And I'd like to add a variant that
operates on byte arrays, and would like to call it concat().

Also clean up the tests a bit.

Test: atest com.android.internal.util.ArrayUtilsTest
BYPASS_INCLUSIVE_LANGUAGE_REASON=existing code in SettingsBackupAgent
Change-Id: I21e958d3d7b3366ec724e757ac4ccc50b36abee1
Merged-In: I21e958d3d7b3366ec724e757ac4ccc50b36abee1
(cherry picked from commit 962323f6e8ed40088a6b5a29fd59b1b39660ef7b)