Make brightness transition to ON immediate if the screen was OFF because of close proxiity
Currently, we ramp the brightness change from the start to end value at
a configurable speed. If this configuration is set to a small value,
users can have a deteriorated experience for the cases when they are on
a phone call with phone on their ear, and then immediately want to
check the phone display. This CL is to fix this experience, and make the
transition immediate for such scenarios

Bug: 238573816
Test: Manual
Change-Id: I08a441641d836fcca746be549d7327e29cf2fb59