Pass keyguard occlude state to onLaunchAnimationCancelled.
We need to apply this canonical occlusion state whenever an
animation that occludes the lockscreen is cancelled. We
usually do this via KeyguardViewMediator's occlusion launch
animator, but the controls/wallet launch cases do not use
this code path, so the WM occlusion value was being ignored.

This is normally fine, but in an edge case where the animation
is cancelled before it begins, this resulted in a temporary
mismatch of the occlusion state, leaving an empty wallpaper
with no lockscreen rendered.

Fixes: 240347798
Test: atest SystemUITests
Test: launch controls activity over lockscreen with a patch to force WM to instantly cancel all remote launch animations
Change-Id: I8a11db3a3b277364c1d8415aa152b352aa8a1c36