Poll for battery changes when Android is interactive
Poll for battery changes for input devices when there is at least one
battery listener and the Android device is actively being used by the
user.

Polling for battery changes is a fallback to ensure that the battery
state for a monitored device is updated regularly, even if the kernel
does not notify userspace about changes to the device's battery. Since
this is only used as a fallback, we only poll for the battery state of
monitored devices once every 5 minutes.

DD: go/inputdevice-battery-notifications

Bug: 243005009
Test: atest InputDeviceBatteryControllerTests
Change-Id: I0ddd019881efeccee07d088bc0e6d6ff9b0d1a54