Don't collapse the shade when SHOW_BOUNCER
It's not necessary to track showing a pending
bouncer to show since showing the bouncer now
wakes up the device. It's also not necessary
to collapse the notification panel when
showing the bouncer from an unlock attempt.
This is also an old code path.

Test: atest SystemUITests
Test: atest BiometricsUnlockControllerTest
Test: MODE_SHOW_BOUNCER from AoD and display off
Test: On a device that supports rear fingerprint
  1. enroll fp
  2. restart device
  3. attempt to authenticate from screen off
  4. see bouncer appears (no regression)
Test:
  1. on device with udfps, expand the shade and
tap on a clickable notification to see the udfps bouncer
  2. Fail udfps 3 times so that the primary bouncer appears
  Observe: the primary bouncer appears instead of going back to LS
Fixes: 240637815

Change-Id: I5fcd53c4942a855d15dc1c11e22995761a69741c
Merged-In: I5fcd53c4942a855d15dc1c11e22995761a69741c