Add the secondary sub to internet dialog
For feature auto data switch:  The feature only
applies to non-DDS. If enabled, we automatically switch to the non-DDS for data traffic when it has better availability than the DDS.

After the change(UI required by b/247880971), when the secondary SIM is being used, it will show up
on the internet dialog and be highlighted. Taping on the gear icon will
bring the user to the secondary SIM setting page.

Test: atest + manual
Bug: 258510998
Change-Id: I41156c29529623e661318caae90edbb9f7d184fe
Merged-In: I41156c29529623e661318caae90edbb9f7d184fe