Changing hsum property in OTA won't alter mode
If a device is set to a particular hsum property in its build files, but
during an OTA that property changes to another value, the device
continues to operate in the previous mode. This is essential otherwise
user 0 will become invalid.

Actual mode is now determined by the system user's base flags, rather
than by the build property, so even if the property changes, the
previous flags will be respected and the mode won't change.

WWARNING: This will not work properly if the pre-OTA build is Q or
earlier (specifically, userVersion < 8). So we are demanding that no Q-
device upgrades to U+ directly while also changing the hsum mode.
Attempting to do so will corrupt the system user.

Bug: 251802540
Test: Manual (flash headed, then flash to default headless and confirm
that it's still headed; and vice versa)

Change-Id: I1b5f79c896de10eab2de1eb133d4f7a61a0c06f5