Revert "libandroidfw: use map for more efficient indexOfString()"
This reverts commit 175aab902a9f48571a20321a86b45fce8e02c757.

Reason for revert: b/253420843

This CL causes memory regression as described in above bug. To speed
up the lookup from resource string to resource ID, the original CL
introduced std::unordered_map for the purpose, unfortunately, that
adds much memory footprint due to the number of key strings per
process. We will need to revisit the solution with a more memory
efficient approach.

Change-Id: I9c6754d361f091e963d4bebea0e698e72e221e51