Fixing QS not showing up in split shade
This is fixing two issues:
- QS invisible when expanding on keyguard from clicking notification
- QS invisible when unfolding device, because then we're effectively keeping shade open when one screen is turning off and transitioning to keyguard

The issue was caused by ag/20007064 because now we rely on qsExpanded state being explicitly set in split shade. Turns out we were only doing that in updateQsExpansionForLockscreenToShadeTransition only until current state was not SHADE_LOCKED but now this state is set before we do the animation, so it was never called. This fix is removing that condition and making it only work when we're in the transition (pxAmount > 0), otherwise state would reset as the transition is reset at the end.

Fixes: 253192346
Fixes: 253978080
Test: on split keyguard, click to expand notification, see QS visible on the left
Change-Id: I6e2d118eeb889d188dcd0a5dc2f21f5bedebf811