Support for unavailable and disabled affordances.
End-to-end implementation (across all layers of both System UI and the
shared library) for supporting a picker state for affordances, allowing
config implementations on the system UI to easily convey whether their
affordance should be hidden on the device or should be marked as
disabled and show a dialog with instructions for re-enablement.

Bug: 256695447,256695925,256695924
Test: unit tests still pass. Was able to see on a tablet that doesn't
support NFC that wallet is gone as an option and, when I deselected all
home devices, the home affordance was disabled with the correct
instruction dialog when picked. Then, marking at least one home device
as favorite, the picker allowed me to select "home" again.

Change-Id: I77637f28cdacecaa02ab33603bbc008a501d03d1
Merged-In: I77637f28cdacecaa02ab33603bbc008a501d03d1