Update BAL Check to return Allow Reason
As part of go/activity-security, we are restricting when activities
can be launched to a subset of those allowed by BAL. This commit
updates the BAL check to return the reason why a BAL launch was
allowed. This will be used according to go/asm-buckets to decide
whether ASM checks are also required.

Note: categories subject to change after more data is gathered.

Bug: 246738504
Test: atest ActivityStarterTests BackgroundActivityLaunchTest

Change-Id: Ie2bb56c53e49f6511a3e830ed1fa228b26e93a05