Fix ACTION_VOLUME_CHANGED behavior
The AudioDeviceVolumeManager.setDeviceVolume() method enables
changing the volume on a device that is not the current device
for a given stream. But the ACTION_VOLUME_CHANGED intent was
fired for any volume update, regardless of the device affected.
Since the intent doesn't indicate for which device the
volume changed, an application registering for this intent
would receive volume updates that don't match the current
volume settings. This change disables firing the intent when
the device is not the current device for the stream type.

Bug: 253118335
Test: atest android.media.audio.cts.AudioManagerTest#testVolumeChangedIntent
Change-Id: I31ca413666e9db94b1ed2db20c87effb46753e3e