BinaryTransparencyService: Update adb shell cmd.
In the default mode, we switched from SHA256 to APK's content
digest to compute package hash.

However, we are preserving the SHA256 scheme as an option for
the cmdline, and this does not have to be super performant, as
long as it is not the default measurement that is used frequently,
since it is computationally intensive.

Bug: 259347186

Test: Manual.
	`adb shell cmd transparency get mba_info -o`
	`adb shell cmd transparency get apex_info -o`
	`adb shell cmd transparency get module_info -o`
      Also, repeat the commands with `--no-headers` option.
Change-Id: I5b8bd42067b7835653c661b4943f08f4e05eecc2