increase HotwordAudioStream copy chunk size
Each read to the read end of the HotwordAudioStream pipe provided by the
HotwordDetectionService is now performed with 32KB size. This request
will return early with whatever bytes are available to be read.

The reason for increasing the default buffer size is to improve the
throughput latency of data egressed from the HotwordDetectionService.

Bug: 262433815
Test: Manual testing with test app
Change-Id: I922bd28e6018bd5abb1c8b5e5e79f500c0b4c093