Adjust concurrency limits based on device.
The concurrency limits haven't really changed much since JobScheduler
was first created. Since we're moving to a world where we want apps to
run jobs instead of FGS, we need to adjust them to handle the new
reality (fewer FGS, more jobs) and make sense given modern device
characteristics. We now set the concurrency limit based on the device's
RAM size and set min and max slots as a percentage of that dynamic
limit.

Bug: 262613827
Test: atest CtsJobSchedulerTestCases:ExpeditedJobTest
Test: atest CtsJobSchedulerTestCases:JobSchedulingTest
Test: atest FrameworksMockingServicesTests:JobConcurrencyManagerTest
Test: atest FrameworksServicesTests:BiasSchedulingTest
Test: atest FrameworksServicesTests:WorkCountTrackerTest
Test: atest FrameworksServicesTests:WorkTypeConfigTest
Change-Id: Ic0aa25f8d2c6af5800f9b33359ac58d1b96435f3