refactor(#AlwaysOnMagnifier): Use real controller instead of mock in MagnificationControllerTest
We change to use a real FullScreenMagnificationController instance in test. Originally, we use mock controller with stub methods. In some cases, it can not reflect the real code behavior. Therefore, we change to use a real controller.

In MagnificationController, we remove the condition check in onRequestMagnificationSpec() method. The onRequestMagnificationSpec() method would be triggered only when the fullscreen magnifier is magnified. In this case, the request doesn't matter and we should always disable the window magnification.

Test: atest MagnificationControllerTest

Change-Id: If390a43b1858296d9050fa3ec909efcf3d378ae9