Send PACKAGE/CANCEL_ENABLE_ROLLBACK as foreground broadcast.
Rollback system expects timely delivery of these broadcasts,
which broadcasting system does not provide. Ideally, these
should be a direct callbacks but sending them as foreground
broadcasts seems to work for now and safe to do so.

Also add an explicit dependency of RollbackTest app for
MultiUserRollbackTest, so that that the app gets rebuilt
if needed when building and running the test.

Bug: 259202331
Test: atest com.android.tests.rollback.host.MultiUserRollbackTest

Change-Id: I6ede9bfd4a5d170fb7448a1417433593fdd9e332