Implement Font Scaling Quick Settings Tile (2/n)
Font scaling quick settings tile provides a quick access for users to
change the Font Size in quick settings panel.

In this modification, we create the FontScalingTile and set the icon and
title for the tile. Only the appearance of the tile is set.

Disabling the tile with setting isAvailable to false.

Bug: 242326166
Test: Attach a video to the bug
Test: atest TilesStatesTextTest QSTileViewImplTest QSFactoryImplTest
Change-Id: I5d4d3ddc97eb1db90db48f23e362136b002d4b2d