Add initial corner radius to pin button.
In master, I removed setting the corner radius in onLayout because it
was being called everytime we would push on a pin button and caused a
janky animation; however this has seemed to cause a regression in the
sim pin view. Only set corner radius on the first onLayout pass.

Fixes: 268297323
Test: open sim pin view.
Change-Id: I3cb2d9fa74cb43c8a10b670b823d44d4f1020b88