Don't trigger links for touches outside of line bounds
LinkMovementMethod currently allows triggering links for touches outside of line bounds. A link on the top line can be triggered by a touch above the top of the layout, and a link at the end of a line can be triggered by a touch to the right of the end of the line.

This changes the behavior to only trigger links for touches within the line bounds.

Bug: 271473145
Test: atest LinkMovementMethodTest
Change-Id: Ib049c231cef08fa5f55160d32cb78eb993de93d9