CSD: Remove a2dp for safe volume index devices
Currently there is no good way to identify whether an a2dp device is a
bluetooth headphone for which the sound dose calculation can be applied
or just a bluetooth speaker (e.g.: car kit). Disabling the CSD
calculation for a2dp for now until we have a good solution for this use
case.

Test: dumpsys audio, audio_flinger
Bug: 276660720
Change-Id: Iaced9137481ff82eaa51ef60e351371195d04a38