Add manifest attr length and tag count checks
There are multiple security vulnerabilites related to a lack of any
restrictions on manifest attribute lengths and tag counts. These
restrictions will now be added to the parser to prevent installation of
any malicious apps with excessivly large string attributes and/or large
number of repeated tags in the manifest.

Bug: 267185315
Test: Manually tried to install malicious apps to check that it fails
install
Test: atest com.android.server.pm.parsing

Change-Id: I00d4c5eea60a8ae01f8620a53412a9f9cfadc1bc