Reduce unnecessary overhead in CpuWakeupStats
Only the wakeups with device mappings to known subsystems can be
meaningfully attributed. Skipping all other wakeups so that they will
no longer be tracked by CpuWakeupStats. These wakeups can still be
found in batterystats history.

Also tweaking some parameters and making them configurable:
- Waking history retention - reduced the default to five minutes. Fixed
a bug which was not removing the history past retention properly.
- Matching window - increased the default to one second.

The recent waking history is not expected to be high volume data, so
keeping it around for the full retention period as it is useful in
debugging and allows more room for arbitrary delays in reporting for the
wakeup to be correctly attributed.

Test: atest \
 FrameworksServicesTests:com.android.server.power.stats.wakeups

Test: Manually verify the output of
`adb shell dumpsys batterystats --wakeups`

Bug: 271496233
Change-Id: Ie6780073bc6898fe9c2c7163074631eb619c1392