Implement InputConnectionWrapper#replaceText()
This is a follow up CL to our previous CL [1], which forgot to
implement

  InputConnectionWrapper#replaceText().

As a result, it has alreays fallen back to the slow path even if the
wrapped InputConnection does have more efficient implementation of
InputConnection#replaceText().

With this CL, wrapped InputConnection can also start taking advantage
of the optimized version of InputConnection#replaceText() if
available.

This CL also adds InputConnectionWrapperTest#verifyAllMethodsWrapped()
to verify that tall the default methods defined in InputConnection are
implemented in InputConnectionWrapper.

 [1]: Ia0e42270cfae112fcf19205a7151e4ab4ac4afdd
      ae2ae5defd4cc2a3de6d2ad8af46c989a39530c2

Bug: 241503197
Fix: 279960381
Fix: 281150084
Test: atest FrameworksCoreTests:InputConnectionWrapperTest
Change-Id: I02a3549f72bf969812bb924a51c18fc81a2668af