fix(#MagSettingsPanel): fix panel position not correct after unfold/rotation
In WindowMagnificationSettings#onConfigurationChanged, for case CONFIG_SCREEN_SIZE/CONFIG_ORIENTATION, we change to directly reset the panel position to the default bottom-right corner. Originally it would try to keep the same Y ratio in bounds, but the new calculated might be not proper in some case. Therefore, we just reset the panel position to the default.

Bug: 279544895
Bug: 279565372
Test: manually
      atest WindowMagnificationSettingsTest
Change-Id: If57f685b3ce8309c74541dd33a0f9b5902d95ae5