InputSettings: fix touchpad tap-to-click default
This was inconsistent with the default in
InputSettingsObserver#updateTouchpadTapToClickEnabled, leading to the
settings UI showing the setting being off on a fresh install, when
actually it was on.

Since the source of this bug was having the default in two different
places that could be inconsistent, also refactor InputSettingsObserver
to use the InputSettings methods where applicable.

Bug: 283135635
Test: wipe the device, check Settings > System > Touchpad > Tap to click
      toggle is shown as on
Change-Id: I8418c35c674a04e6849bd05de9e3f37a793e3aa9