Consider process proc_state when computing runnableAt.
It is possible for a process to be in a particular
proc_state even if the uid that the process belongs to
is not in that state. We could modify the uid observer
to consider all proc_states below TOP as foreground
but compared to it, this change is less risky.

Bug: 287015744
Test: atest services/tests/mockingservicestests/src/com/android/server/am/BroadcastQueueTest.java
Test: atest services/tests/mockingservicestests/src/com/android/server/am/BroadcastQueueModernImplTest.java
Change-Id: I0df7052244bdbee0342a2af5ab5da3c11cd8c86b