Show the overflow expanded view in bubble bar
When the overflow bubble is selected in the bubble bar
we now show the expanded view for it. Rather than calculating
the Y for expanded views and align them with the top,
the positioner now calculates the bottom of the expanded view
so that we can align them above the bubble bar easily.

Fixes: 287121592
Fixes: 271468319
Test: - atest BubblesTest
      - atest BubbleOverflowTest
      - manual:
        - Enalbe bubble bar
        - Add a bubble
        - Expand the overflow
        - Expanded view should show the overflow state
Change-Id: I2db22a8e7878b5c644d72472e14848b0d36a62b3