Propagate channel updates to their conversation children
When updateNotificationChannel is called for a channel, these changes are propagated to the conversation channels of which the one being updated is parent. This is done by applying the "diff" of the update to each child, only to fields that are not "locked" (i.e. previously set by the user).

Flag: PROPAGATE_CHANNEL_UPDATES_TO_CONVERSATIONS
Test: atest PreferenceHelperTest + manual
Bug: 285932635
Change-Id: I1cf2905ec861d82e68a0a7483838a9b8ca5fdb19