Handle when sim state indicates that it is absent
We want to update the sim state the view mediator when the sim state
indicates that it is absent. When we disable the esim, we want to be
able to update the sim state to indicate that we no longer have secure
sim pin when we swipe to unlock. Additionally, we want to ensure that
mPendingPinLock is set to false so that we do not show keyguard again
after dismissing.

Fixes: 286520191
Test: disable esim in sim when security method is None
Test: unlock sim when security method is None
Test: disable esim in simpuk when security method is None
Test: unlock simpuk when securith method is None
Test: disable esim in sim when security method is Swipe
Test: unlock sim when security method is Swipe
Test: disable esim in simpuk when security method is Swipe
Test: unlock simpuk when securith method is Swipe
Test: disable esim in sim when security method is Pin
Test: unlock sim when security method is Pin
Test: disable esim in simpuk when security method is Pin
Test: unlock simpuk when securith method is Pin
(cherry picked from commit 5866a4a23a1ce4497b04a58133201363f5ea5686)
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:9c10ce1a928fcdae1db572bdbd97c40128593cb1)
Merged-In: I69ef1698d1c3b3833cbbe925999dd4aa87cc00eb
Change-Id: I69ef1698d1c3b3833cbbe925999dd4aa87cc00eb