Fix media being stretched during the wake-up/unlock animation
This happened only on large screens/split shade, and when there was an
explicit delay to play the doze wake up animation.

During this delay, the device will not be in the "dozing" state anymore,
therefore media was showing up, but the rest of the keyguard was still
positioned as if it was in the "dozing" state, centered, and using the
full width of the screen. Since media is a child view of keyguard on
large screens, it was also using the full width of the screen.

The fix is to keep media hidden during this delay.

Fixes: 283479104
Test: KeyguardMediaControllerTest.kt
Test: NotificationPanelViewControllerTest.java
(cherry picked from commit 574367ed8158156c6a17d4ca260a0651dcdfbc04)
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:38eeb15a3b8807363d4a777aa9f540fb4e0ec87c)
Merged-In: Id004a74d95e66e7f33ef52aca0dc655db767b893
Change-Id: Id004a74d95e66e7f33ef52aca0dc655db767b893