Log plug state chage & other events at boot
Logs plugged state changed, charging state changed, and battery level to
statsd at boot. These events used to be logged at boot, but the logic
was changed in ag/19715391. This is a partial revert of ag/19715391 to
restore the logic to the old behavior.

Bug: 289301086
Test: manually verified the atoms were logged at boot
Change-Id: Ia5136531ec4fa1014b44bcf2701af543b66bfe45
Merged-In: Ia5136531ec4fa1014b44bcf2701af543b66bfe45