chore(#Magnification): use persisted scale if zoom-in-temporary when magnifier is zoomed out by service
Originally when the fullscreen magnifier is activated, using triple-tap-and-hold to temporarily zoom in would use (current scale + 1) as the furthur zoom in scale.

We add two cases that would zoom in with persisted scale.
1. User just tap the shortcut to activate the magnifier.
2. The magnifier just zooms out due to user context changed such as transitions or rotations.

For these two cases we conclude the expected bahaviour should be zooming in with persisted scale.

Bug: 289131167
Test: manually
      atest FullScreenMagnificationControllerTest
      atest FullScreenMagnificationGestureHandlerTest
Change-Id: I0a48e1465763c5387bbac537471de82efff7caf5