Attempt to reduce flakiness of testTaskChangeCallBacks
The test requires the device to be in a non-sleep state. Such as
notifyTaskRemovalStarted is only called when a resumed activity
is finishing.

Sometimes SystemUI was crashed on other test module, then keyguard
is shown after SystemUI is restarted. That causes the test activities
to be unable to keep resumed state so the expected callback is missed.

Bug: 291149606
Test: atest TaskStackChangedListenerTest#testTaskChangeCallBacks
Change-Id: I5f512c548f0bf28a344e8cfab4e08c9fee5e7221