Clarify reasons for notification removal.
The removed list of reasons is incomplete and might make developer think
that a notification can't be removed for any other reason. "reason"
argument in one of the overloads provides a link to full list of
reasons.

Annotating the argument with @NotificationCancelReason generates an
equivalent of @param javadoc with the complete list of reasons.

Bug: 292204120
Test: doc-only change
Change-Id: I1394b987e8ccdab474189f5fbb57ea3c70bce50b