Removing system bar insets from the floating menu layoutParams.
This will allow our shortcut buttons to display on top of the navigation bar, rather than being cut off.
Additional adjustments are made to ensure proper placement in different rotations, as well as use of cutouts to maintain visibility when tucked.

This change also establishes a feature flags file for SystemUi Accessibility.

Bug: 283768342
Test: adb sync system_ext && adb shell stop && adb shell start
Change-Id: I5206a6a1433597ca35aa66dc8bb5cf76d29a35bc