Don't skip resume on transient-launch activity
When transient-launching, we are trying not to affect any
other lifecycles by pausing things. This should imply that we
shouldn't wait for ongoing pauses since they are independent
of the transient-launch.

deferPause is used to indicate that the current resume won't
pause/wait for the current activity to pause. So, don't wait
for ongoing pauses that started beforehand either.

Bug: 294152870
Bug: 293379641
Test: have a decoupled slow-pausing app's pause (eg. CCT) be
      interrupted by recents entry.
Change-Id: I4a9b44f1f08bfd2e41eee9352588b84ee8cc8a5d
(cherry picked from commit 956baab0fb1ce40268b9f98bf6347f71b86de75b)