Fix bugs revolving around user switching.
When switching users in the bouncer, we do no invoke
startAppearAnimation. Additionally, the theme color change was delayed
for roughly 3 seconds after the user switch completes, which reinflates
the bouncer view.

In order to handle this case, I've moved some logic from
startAppearAnimation to onViewAttached. Additionally, when the user
switch occurs, we opt to reinflate the bouncer. When the color changes,
we do not reinflate the view. Although the view will not adopt the new
theme changes, it will change when we leave the bouncer and enter again.

Fixes: 292255597
Fixes: 294035459
Fixes: 292123727
Test: switch users from auto confirmation pin to pin and then back.
Test: switch users from locked out pattern to pin and then back.
Test: switch users from password to pin and then back.
Change-Id: I5b851598c09a0587dd241838b4af256b6c533aaa