Show volume UI when adjusting volume with direction ADJUST_SAME
When HdmiControlService is notified by AudioService of an audio
adjustment with direction ADJUST_SAME, it sends a <Give Audio Status>
message and notifies AudioService of the response, causing it to show volume UI.

This is handled by AbsoluteVolumeAudioStatusAction. It requires
adding a boolean flag to override the default behavior of NOT notifying
AudioService of an updated audio status, if the updated audio status
didn't contain a new volume or mute state.

Bug: 293445117
Test: atest TvToAudioSystemAvbTest; manual
Change-Id: I3dc857c569f587f52f3f2f80446e84b27ad8abda