Avoid using the same previewMode logic between lockscreen and WPP
Previously `previewMode` lived inside `KeyguardInteractor` which is a SysuiSingleton.
This caused issues as even though we had different copies of `KeyguardRootView` for lockscreen and WallpaperPicker, both Views were using the same value of previewMode, which is never explicitly disabled upon leaving WPP.

Split up the enabling of previewMode between the KeyguardRootView and the Quick affordances, as the KeyguardRootView doesn't need to know about the quick affordances.

Fixes: 296249943
Test: atest KeyguardQuickAffordancesCombinedViewModel.kt
Test: atest KeyguardRootViewModelTest.kt
Test: manual -- go to WPP and shortcuts customization, leave, go back to lockscreen. The UI is now different between them as intended
Change-Id: I81a059002e179bb80be40617ce427d595b58b616