Set a view id for the bubble dismiss view (shell)
There is a flake in the drag to dismiss bubble test, I think it's
because sometimes it misses the dismiss target since the test drags
the bubble to the bottom of the screen instead of the target location.

In attempt to fix this I'm setting an ID on the dismiss target to
look up where it is on screen and drag the bubble to it directly.

Test: treehugger
Bug: 296933279
Change-Id: I6c7040cc05d008604b863b066036b4e06e2dfc72