Truncate user data to a limit of 500 characters
Fix vulnerability that allows creating users with no restrictions. This is done by creating an intent to create a user and putting extras that are too long to be serialized. It causes IOException and the restrictions are not written in the file.

By truncating the string values when writing them to the file, we ensure that the exception does not happen and it can be recorded correctly.

Bug: 293602317
Test: install app provided in the bug, open app and click add. Check logcat to see there is no more IOException. Reboot the device by either opening User details page or running adb shell dumpsys user | grep -A12 heen and see that the restrictions are in place.
Change-Id: I633dc10974a64ef2abd07e67ff2d209847129989
Merged-In: I633dc10974a64ef2abd07e67ff2d209847129989
(cherry picked from commit 16d9c94f262ea1a25ebac02486a2e2ada832df6a)