Apply default policy for enabled rules without custom policy
When an automatic zen rule doesn't have a custom policy set, we were not applying the default rule until the end to fill in any settings that were unset by any other rule's specific policy. This means that when an enabled rule used the default policy, it would not apply even if the merged consolidated policy would otherwise have kept the default policy's setting due to it being stricter.

Fixes: 297415089
Test: manual, ZenModeHelperTest
Change-Id: I641fdb53f023c0b37b5871550e59962717ef529c