Use Task Position to check if task is below disallowed area
Currently, the TaskPositioner uses the raw y position of the
motion event to check if the task positioner should finish handling the
drag event by calling WCT#setBounds or if it should allow something else
handle it (i.e. DesktopTasksController handling it by transitioning the
task into fullscreen). Meanwhile, DesktopTasksController uses the
position the task is placed by the TaskPositioner (which is calculated
by adding the distance of a drag to the original position of a task) to
check if the visual indicator should be created/destroyed. These
differences cause the visual indicator to be shown in positions where
the task may not transition into fullscreen when released.

This change uses the same method used to position the task (adding the
distance of the drag to the original task position) to check if the
TaskPositioner should handle the end of the drag.

Bug: 296924625
Test: Drag freeform task to top and release as soon as visual indicator
appears. Task should transition to fullscreen.

Change-Id: I80777ee77b502e9e4bcfe40c3a5bed5139758e90