Post UserTracker callbacks to background
UserTracker has been blocking binder thread while notifying the
callbacks. We expect this process to be asynchronious. It's impossible
to reuse current implementations because there is problem with a deadlock if the callbacks processing runs on the same thread as the callback (which is blocked by the latch). Before the change it works because the callbacks are processed on a binder thread which is always distinct. This is the issue that this feature addresses.

Test: atest UserTrackerImplTest
Test: atest OverviewProxyServiceTest
Test: atest KeyguardUpdateMonitorTest
Test: atest UserTrackerImplReceiveTest
Fixes: 283101304
Change-Id: I47cbc755f1be0025aec00f5ff1ee419cf5a06a29