[flexiglass] SceneContainerFlags.
This is a new injectable utility class to replace FeatureFlags(Classic)
for Flexiglass specifically. It abstracts away the need to check the
various feature flags, compose availability, etc and reduces it all to a
single isEnabled call.

Bug: 283121968
Test: added new unit test
Test: all existing unit tests pass
Test: manually verified that Flexiglass correctly runs when all
conditions are met
Test: manually verified that Flexiglass correctly doesn't run when any
of its conditions are unmet: compose is off, the main feature flag is
off, or any one of the dependnt sub-flags is off.

Change-Id: I8b8b1b6a761c6bac6e32ba65ec76af1e7964cb6c