Only draw legacy nav bar background when it is visible
We checked the nav bar visibility before the commit:
3a5e13e237115e33d10aca4bb4c2edc26a698f28 .
But it was based on mNavigationColorViewState.visible which would be
always false if FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS is not applied. So the
commit removed the visible check, and caused the black bar issue on
dream windows.

This CL adds the visibility check back, uses the requested visibility
instead, which is independent from FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS.

Fix: 273495037
Test: Enable screen saver, and press power putton to show it while in
      split-screen mode. Make sure there is no black bar at the bottom.
Test: Show IME in combinations of
      a. floating mode or non-floating mode
      b. gesture navigation or 3-button navigation
      c. taskbar or navigation bar
      d. light theme or dark theme
      And see if there is anything unexpected
Change-Id: I9b44aa801da7d5be2938406b47b639616f373e43