Fix media carousel activities in RTL
This CL fixes the change in reordering, players visibility, and media
scroll handler issues in RTL.
On reordering all players, the visible player should be the rightmost player.
On playing a UMO View, onLayout method is called. It adjust the scroll
value in RTL when the view is not laid out. Therefore, we need to set
the scroll value to the relative value.
On removing a player, we need to make sure the scroll view changes the
scroll value if we are on the rightmost player in media carousel.

Bug: 289174065
Bug: 287168003
Test: Manually checked the media carousel using 3 different umos at the
same time.
Test: atest MediaCarouselScrollHandlerTest.

Change-Id: I80a32eab50d3b75cf3e617d7c8e8d961b2eb6b8a