Avoid crashing system if the installed user doesn't existing
It is possible the user may be removed during the package install,
and it may happen frequently in the auto testing. This may cause
the system crash. Instead of making system crash, we add an user
checking in the code that may crash the system to report error to
the caller.

It's good to get notified when the user is removed and the util can
do a better handle here. We will try to go through in the follow up
bug later.

Bug: 301203319
Test: manual. Comment the install shell command user checking code
and install package on invalid user.

Change-Id: I168fe55d4a0604ec86f502ac8e434407739478ea