Add burn-in for shade migration
This migrates burn-in logic away from KeyguardClockPositionAlgorithm
to flows. Does so by using a constraintlayout Layer, in order to group
elements for scaling and translation.

Weather clocks scales, but smartspace is misplaced and will follow up
work is required on the large version.

Aod notif icon translation has been disabled, as it interferes with other
translation. It is no longer needed in latest motion specs.

Fixes: 296372316
Test: atest KeyguardRootViewModelTest BurnInInteractorTest
Change-Id: Iebc5ca66d441e363d61c3cc536e3a5ae4d0d962e