Refactor Handle Menu as Single Window
Refactored the handle menu setup code so it is inflated as a singular window rather than 3 different windows. This will allow for much easier animation and shadow addition. Divided the handle menu setup function into 3 (one for each pill) for readability. Also added 1dp shadow with elevation. Also removed the close button as it is not needed in freeform or fullscreen modes.

Test: Manually tested in freeform and fullscreen modes
Bug: 303285820,301493992
Change-Id: I4c02748b76ff3089efb5e4a17e20c9260f81b1cb