Fix size caching bug in (Horizontal|Vertical)Grid
This CL fixes a caching bug in (Horizontal|Vertical)Grid that would
occur when the size of an element in the grid changed size. We would
keep only the maximum size the element ever had into accout because we
never reset the values in the size cache.

See b/291053742#comment12 for before/after videos.

Bug: 291053742
Test: Manual, see b/291053742#comment12
Change-Id: I6db5bea7cf8012e39745f94ecbce5411275f37c2