Refactor processing cred response
A single credential request is requested from gathering multiple
credential options. Then we have been processing the response as if it
was a single request/response. In reality, the request is more like a
batched request and the response is a batched response. By mapping
provider info list to autofill id, we can produce a set of provider
info list. Each provider list will be an indenpendent response of the
batched response. Then we can process the provider info list for each
autofill id independently, sorting them and applying inline max count
logic.

Test: local device testing
Bug: 299321128
Change-Id: Id83b1d5e5c4f6d66da9259c3943502c3ec8a48bb