Make fullscreen caption the size of the caption handle
The current fullscreen expands the full width of the window even tho it
is the same color of the status bar and only needs to show the handle
bar. To reduce the load on the composer, this change shrinks the size of
the caption to the width of the handle bar.

This change also removes RelayoutParams#mCaptionX and
RelayoutParams#mCaptionY since they are used but never assigned.

Bug: 305983723
Test: Manually opening/closing and dragging from the handle bar and
caption

Change-Id: I4f6fc95425486a9699455ee0c49d40d2f50e5e27