Switch pin_shared_anon_size namespace
Use the `runtime_native` namespace instead of `runtime_native_boot` to
ensure the committed flags align with dynamic runtime behavior. We want
this flag to be responsive to post-boot flag updates to maximize the
responsiveness of experiments, both for collection and in the event that
any issues arise.

Note that `_native` is used as we reference the flag during system
server initialization.

Test: atest PinnerServiceTest
Bug: 309803726
Change-Id: Ic142b9ccf75015d386999aeb45e922fc903f2cb4