Rename the fix_unlocked_device_required_keys flag
This flag was defined as a regular flag and then was later changed to a
fixed_read_only flag.  This scenario is currently "unsupported" by the
flags infrastructure; an error occurs when trying to advance the flag to
staging.  Work around this by renaming the flag so that the flags
infrastructure sees it as an entirely new flag.

(Also fix two formatting errors in LockStateTrackingRule.kt which were
causing the ktlint_hook pre-upload hook to fail.)

Bug: 296464083
Bug: 311648623
Test: build
Change-Id: Ia8c7b3d2602bcbd4e349cc7f340ff6077412996e