Fix legacy udfps occluded state transitions
In the udfps legacy code, add support for
occluded <=> AOD transitions which were
previously unsupported and would put the udfps UI
in the wrong state.

Also fix the transition from AOD => Occluded after a cancellation
of a AOD => LS transition.
For example the following sequence of transitions when the user
is double pressing the power button to show the camera:
  START: AOD => Lockscreen (first power button press)
  CANCEL: AOD => Lockscreen (second power button press)
  START: Lockscreen => Occluded
In this case, ensure the doze amount is appropriately updated to 0f.

Test: atest SystemUITests
Test: atest UdfpsKeyguardViewLegacyControllerWithCoroutinesTest
Test: Show the camera app over the lockscreen & then transition
to AOD. Observe that the AOD udfps icon shows, not the LS version
Fixes: 313963031
Fixes: 313954401
Flag: NONE
Change-Id: I2b3c3759001dd51d4374456511d27f5552ec0b76

Change-Id: I055fee4ea35420211b6e68eb77130ab47864e3a0