Deprecate ShadeExpansionStateManager
The original goal was to delete ShadeExpansionStateManager, but at
this point, that would require more effort and regression risk than
it's worth. The remaining usage of ShadeExpansionStateManager is in
code that will naturally be migrated or deleted by the Flexiglass
project, so spending effort to fully delete it now would be
wasteful. It is better to simply deprecate and move on.

Fixes: 269323368
Test: N/A
Flag: N/A
Change-Id: I159e8c536a2f243b0d96099a3078ab751e1e658c