Deprecate InputMethodSettings#switchCurrentUser()
Previously we have reused the same instance of

  InputMethodManagerService#mSettings

across users, which needs to be updated before supporting concurrent
multiple users in InputMethodManagerService.

With this CL a new instance of InputMethodSettings will be recreated
every time the current IME user is switching.  This is an important
milestone to keep maintaining multiple instances of
InputMethodSettings for each user.

There must be no observable behavior change in this CL.

Bug: 309837937
Fix: 309868254
Test: presubmit
Change-Id: I52297912d0b66e5e40b6a624c35427a377dc7b1b