Formalize BlockedNumberContract.SystemContract APIs
Formalize the BlockedNumberContract.SystemContract APIs as part of the
Telecom modularization. There are several hidden APIs that are being
referenced from Telecom that need to be resolved. The
blockStatusToString() implementation is being copied over to Telecom
while the other dependencies are being formalized into system APIs as
part of BlockedNumberContract (the hidden references have been
unchanged).

Bug: 308208071
Bug: 311773409
Test: atest BlockedNumberContractTest
Change-Id: I79db751d494e81c9f638a300f871fed021b76808