[ECM] Make flag fixed (and rename it)
Make the enhanced_confirmation_mode_apis flag fixed (a.k.a. read
only). We'll need to use the flag across both framework and the
Permission module, and the only (?) way to support this is to use a
fixed flag.

Also, rename the flag. This is because feature flag documentation
shows the following warning:

    IMPORTANT: Modifying the type of an existing flag is not supported.
    For example, you cannot change a regular flag to a fixed read-only
    flag, or vice versa. If you need to change the flag type, the best
    practice is to create a new flag.

Bug: 316181485
Test: manual
Change-Id: I507cc4c3feae33c7946046b1734230a11cdc5927