Directly propagate progress if setReadyToHandleTransition is called from the correct thread
Before this change, setReadyToHandleTransition was always posting hte callback to propagate progress to the progressHandler. This caused some reordering of callbacks that led to some transition progress to be received out of order.

Now if setReadyToHandle transition is called from the correct thread (the same as progressHandler), the posting is skipped.

Flag: None
Bug: 311084698
Test: ScopedUnfoldTransitionProgressProviderTest
Change-Id: Ib1c596838b7582cef0cee4b4363f74edc77f8c0e