Revert^2 "[Sb] SingleBindableIcon implementation"
109bd7d365f828057415796c484d2f4ec1e33619

Original commit message:
[Sb] SingleBindableIcon implementation

Small subclass of ModernStatusBarView that provides a default binding so
we can keep the mechanics of the old status bar layout logic away from
newer icon bindings.

The usecase for this class is basically: I want to show an icon that can
potentially update based on the state it gets from its view model.

Follow-up CL has the only current client of this. Note that this is not
flag-controlled because there is no code that uses it yet.

Test: SingleBindableStatusBarIconViewTest
Bug: 311417356
Flag: NONE

Change-Id: Id6191d182f79f04ff2ac41aecfd49e6d26e24d70