Fix: insert mode crash when delete
When the editor has a Span that implement UpdateLayout, deleting
text in insert mode will crash the APP. This is due to the fact
that updating the text will also trigger span update. And then trigger
the DynamicLayout updates and display list update. Because the insert
mode applied a transformation on the text, it's hard to know the updated
text range on the transformed text.
This CL fixed the issue by always updating the entire when span is
updated.

Bug: 314254153
Test: mannually tested as the Editor.java is hidden
Change-Id: If4953de0ae37f92683e5c12c7a11b29a43260fa3