[SettingsProvider] move handler to background thread
SettingsRegistry's handler takes care of notifying settings changes to
listeners as well as scheduling jobs. Previously it runs on the main
looper of SettingsProvider, which is effectively on the main thread of
SystemServer because SettingsProvider shares the same process as
SystemServer. Its tasks can be blocking which can block the main thread
of SystemServer. Moving the handler to a background thread instead.

BUG: 319046998
Test: presubmit

Change-Id: I68f022437d1180fdb41d1472524b7e2c277703cb