Clarify InputMethodSettings doesn't require ImfLock
With my previous CL [1], all the internal fields InputMethodSettings
do not need to be guared with ImfLock.

Thus we can safely remove "Locked" suffix from its methods.

Note that the returned values are still subject to change if other
threads are updating relevant SecureSettings, but it's not a new
problem.

 [1]: I512519ed83b28adb2878945bc3e6eaff9ba29f5e
      b624973dc3e82697b17e0633e5dad2f5b212b28b

Bug: 309837937
Test: atest CtsInputMethodTestCases
Test: atest FrameworksServicesTests:InputMethodSettingsTest
Test: atest FrameworksServicesTests:InputMethodManagerServiceTests
Test: atest FrameworksInputMethodSystemServerTests
Change-Id: If370dc0db41c2a6028b1e4f3f58959ccc4d71617