Fix test flakiness.
The default mocked value (0) was causing some code to run that would
occasionally interfere with tests, making them flaky. Setting the value
to the right "unknown" value disables that and prevents the test flakes.

Also add a safety check and escape valve in case something was missed
and things go wrong in production.

Bug: 236261941
Test: atest --rerun-until-failure 25 FrameworksMockingServicesTests:FlexibilityControllerTest
Change-Id: I5d14329e5ba656af77ccdbd39d28d53080c0c47d