Decouple BubbleController from more classes
This change creates BubbleExpandedViewManager and BubbleTaskViewFactory interfaces.

These interfaces allow decoupling BubbleController from more classes so it's easier to write unit tests for Bubble classes.

After this change it's possible to create a unit test for BubbleStackView that does not depend on BubbleController, but there is a blocking robolectric issue b/323188766, so I will add the test once that issue is resolved.

Flag: NA
Bug: 321265700
Test: atest BubblesTest
Test: atest BubbleViewInfoTest
Test: atest BubbleOverflowTest
Test: manual -- build and play with bubbles
Change-Id: Id1c79643c2ad9e88173d6c25a9e88643afaf5564