Fix bubble manage menu after dismissing bubble
After dragging to dismiss a bubble and tapping on the manage button, the manage menu is now displayed correctly on top of the expanded view.

Previously, the expanded view is marked temporarily hidden when the drag motion starts. The surface z ordering is reset at the start of the expanded view animation, and then set again when it ends.
However, because the expanded view is temporarily hidden, the z ordering doesn't get set at the end. This change sets it when the bubble switch animation ends right after the expanded view is no longer temporarily hidden.

Flag: NA
Fixes: 280355857
Test: Manual
      - Add 2 bubbles
      - Expand bubble stack
      - Drag the selected bubble to dismiss
      - The second bubble should be selected
      - Tap on the manage button
      - Observe that the manage menu is drawn on top of the expanded view
Change-Id: I4410c591e88881d9e434dbdb7f7d2e250c0f718b