Hide window immediately if itself doesn't run hide animation
The condition was overextended in commit 563e937 which checks if the
parent container of the window is animating. That causes the window to
wait for animation finish to update visibility, but the animation
finish callback won't happen because itself is not animating. Then the
window that should be hidden remains on screen.

Bug: 302431573
Test: atest WindowStateTests#testIsOnScreen_hiddenByPolicy
Change-Id: Iafc2b2c2a24d8fc8d147354ef2f0b4afeeb510c5