Human-readable XML on Ravenwood.
We had temporarily been using the binary XML implementation in all
situations, but this change brings in the existing `kxml2-android`
library to offer reading/writing of human-readable XML.  We also
use the upstream SAX parser to offer an event-based parser.

In all cases we carefully keep the existing XML libraries in place
untouched on physical devices to avoid regressions.

Bug: 324417456
Test: atest CtsUtilTestCasesRavenwood
Test: atest FrameworksCoreTestsRavenwood
Test: atest FrameworksUtilTestsRavenwood
Change-Id: I2effc7fca5c586968658073f9652d13cf50a2c66