Move UserActionDistance to TransitionSpec
This CL moves UserActionDistance from UserActionResult to
TransitionSpec. I realized while adding custom distances to the demo
that this made more sense: usually the distance is the same when going
from A => B and B => A, and the previous API would require us to specify
the distance twice. Moreover, it was impossible to assign a distance to
a generic transition (for example from any scene to Shade or from Shade
to any scene).

Bug: 308961608
Test: SwipeToSceneTest
Flag: N/A
Change-Id: I24f9cf346f37d81dc19f738665b872e2efcd94d9