Collect the new root task as existence change
Collect the new root pinned task as existence change
to when entering PiP to make sure that it has the mode
OPEN instead of TO_FRONT.

As a side effect, this also helps avoid
the reset of crop in startTransaction
when running Transitions#setupStartState() on the Shell side.

Currently, only making this change under the PiP2 flag
since existing PiP1 leash manipulations via startTransaction
are relying on the current setupStartState flow.

Bug: 325481148
Test: manually-enter PiP in PiP2 and check the transitions logs

Change-Id: I5e2ea95d55822917fbde9857665b8ffa8f8a69ca