Suppress HiddenAbstractMethod on Notification.Style.areNotificationsVisiblyDifferent
The `Notification.Style` class is an abstract class with a publicly
visible constructor and so the assumption is it is expected to be
extended by Apps. As such it cannot have any hidden abstract methods as
that would prevent an App from extending the class. Unfortunately,
`areNotificationsVisiblyDifferent(Style)` is both hidden and abstract.

The issue was not previously detected due to a bug in Metalava but that
has been fixed and so now this is an issue. This change is temporarily
suppressing the error in the code to unblock Metalava changes being
merged into AOSP but it should be fixed properly ASAP.

A proper fix would either require marking the public constructor as
`@removed` or unhiding the method or giving it a concrete
implementation.

Bug: 324468829
Test: m checkapi android-non-updatable-doc-stubs
(cherry picked from https://android-review.googlesource.com/q/commit:ddc69d098f03fce1bed20144f92b8743eddb1767)
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:5387ad207cd9c6686e3aacb93a3ba756d073dad5)
Merged-In: I3b72ef4f0afff35771ee6f37130d1084071e2de1
Change-Id: I3b72ef4f0afff35771ee6f37130d1084071e2de1