[Flexiglas] Split shade - domain layer
Domain layer changes to support the split shade.

Split shade is the single panel shade that's shown for large screens and
unfolded devices where quick settings is one the left and notifications
is on the right.

The most interesting part of this CL is the use of the new
ShadeStartable to hydrate the value of isSplitShade in the repository.

Bug: 328473018
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT
Test: unit tests, integration tests - included with this CL
Test: see followup CLs on this chain for manual testing details
Change-Id: I1952c3b6d392e4d7e3ff71990e23530141821560