Made keyguard pattern view unfocusable
Keyguard pattern view was receiving focus, causing a gray rectangular
background to show up when a physical keyboard was
connected, which isn't intended behavior as physical keyboards cannot
and should not interact with the pattern view.

Fixes: 328650320
Test: Manual - set lock screen auth method to pattern, wake up tablet
with spacebar/enter, ensure that there is no gray rectangle over the
pattern view.
Flag: NA

Change-Id: I768174728eeb269b04d927c25535d586b9a2a6d8