Change how ranking time is calculated
The calculation about 'did the dev supply a "when" value is many years
out of date, and we shouldn't ignore all updates in the name of
stability - updates that a user would consider something new (e.g. sound
or new text) should have a chance to be ranked higher. This shares the
same logic as the notification counts we show in settings

Test: atest FrameworksUiServicesTests
Bug: 326016985

Change-Id: Ib79c1aa903442eaf02ba459d4c399b289d32f91b