Allow to collect moving back task in open transition
Simulation:
 Task A starts(NEW_TASK) Task B, finishes
 Task B starts(NEW_TASK) Task C
   Task B calls moveTaskToBack in onPause
 Task C (declare as translucent and sleep 200ms in
         onCreate to delay transition)
   starts(NEW_TASK) Task B -> attempt to move task B to front
   finishes

Because the initial open transition may be done until the last
step, then moveTaskToBack for B will run after the last step.
Then expect B should be on top but the result is on home screen
because A and C are finished, and B is moved to back.

This change is kinda a workaround to mitigate the case of
deferred move-to-back. Because it is regular to have closing
targets in an open transition, the moving task can run with
the collecting transition directly.

Fix: 325645759
Test: atest TransitionTests#testDeferredMoveTaskToBack
Merged-In: I46d8e26110ebefea74e5dd06c85c003b12693416
Change-Id: I46d8e26110ebefea74e5dd06c85c003b12693416
(cherry picked from commit b105df45e01927b1b3eff1bdfb39e1b3780d1003)