Setting alternate_expand_target to not be focusable
We don't want to make it focusable as it adds one unnecessary step while traversing notifications with keyboard.
It's made focusable implicitly when it's set to clickable by setting OnClickListener in a few places.
By setting focusable to false explicitly, even when made clickable it still stays not focusable.

Fixes: 332702537
Flag: None
Test: Traverse shade with keyboard and see focus doesn't land on notification icon
Change-Id: I78f926524e4192bc1b3954361e9fc564814f8069