Prepare Overlays for MainUser earlier
Headless System User Mode (HSUM) creates and switches to the MainUser
immediately in the initial bootup. This requires that the MainUser's
overlays be ready before launching SetupWizard (SUW).

We therefore monitor for when the user is created, and once it is, we
immediately start updating its overlays, so that they can be ready
early.

Bug: 283904547
Test: Manually confirmed that SUW had theming on HSUM device
Change-Id: Ia3339e6a39bc5d641644cfa6a2a723dd0c067d6f