Replace IMMS#mSettings with IMMS#mCurrentUserId (2nd try)
This is the 2nd attempt of submitting my previous CLs [1][2], which got
reverted due to a boottime regression (Bug 329414077) [3][4]. We have
already addressed the major contributor of the boot time regression [5],
so this CL should be ready to be submitted again.

This CL removes

  InputMethodManagerService#mSettings

as the same instance can be obtained from

  InputMethodSettingsRepository

as long as we keep track of the current user ID.

This is a mechanical refactoring CL. There must be no user observable
behavior change.

 [1]: Idae0fb6d1a9182efe917f81a1d881d488a1b57e5
      b6cafdfd695172cb661b0edf54dcd6b05c6602b8
 [2]: I4b7dc1b0e52e9bf25095eb2e3cc520d46b3869e5
      b7a5ac4370dbbd8612c61eb6a34752afd6b0dbf6
 [3]: I0d65cb9edcdbdaea45b1fa80abfe917b72fb9a50
      76652b0b9d5040c645d379369160e972fcc090ba
 [4]: I585e5c6745c40e51f1080b78367175990b2c5884
      6c90d04779e3c71eb64e995315ef8c66f203759b
 [5]: I8e51b1ced4dc16cdca7e898885c64793665fafef
      f4cc1a7b972f50fac91a5003445c9391f75b0f7c

Bug: 305849394
Fix: 329163064
Fix: 333591545
Test: presubmit
Change-Id: Iaa7f08c9081aac9abe0c95d522185632d575d6b8