Send broadcast to the correct user in QS restore
For broadcast sent to SystemUI (QS restore), use the correct user that
is being restored, both for the broadcast and for retrieving the value
before restore.

This is done only for these broadcasts as the consumers are known and
are expecting this. For the general android package case, an audit is
needed to determine that all consumers are expecting it.

Test: manual, on hsum restore
Fixes: 330304534
Change-Id: Iff392afb96ace3ac9e9223400677e91be71d80f1