Explicitly request to support package restart query
Explicitly request to support package restart query for services
that need to handle onHandleForceStop(). That makes PackageMonitor
MUST register the broadcast receiver for ACTION_QUERY_PACKAGE_RESTART.
Currently, the default constructor is the same as the default
constructor with true. All PackageMonitor will register a receiver.
Behavior is not a change in this CL. There will be a follow up CL to
change default behavior to false, then PackageMonitor only registers
a receiver for service that explicitly requests to support package
restart query.

Bug: 300309050
Test: atest FrameworksCorePackageMonitorTests
Change-Id: Iefe04a0ae594722f793422b23db09d12dc4b05f1