Flag config for bug fix: Talkback focus order
Talkback focus doesn’t move to the “If you change your Google Account picture…” after swiping next to move the focus from “Choose a picture”

The fix require more research as it is not related to focus order in jetpack compose, but rather accessibility modifiers. The group with text was marked as isTraversalGroup.

Bug: 330835921
Test: manual
Change-Id: I83f3ad9273be860b52a412558c1726d80b25710c