[Sat] Use new TelephonyCallback instead of ServiceState
Previously, we were using `ServiceState.isUsingNonTerrestrialNetwork`
to know if a given mobile connection was using satellite
(non-terrestrial). This single bit was insufficient though, given that
there are carrier requirements for hysteresis when losing the
connection.

This new TelephonyCallback.CarrierRoamingNtnModeListener surfaces the
same information as `isUsingNonTerrestrialNetwork`, but it also embeds
the hysteresis information along with it. So we should be able to use
this single bit instead of modeling the hysteresis ourselves.

Test: tests in statusbar/pipeline/mobile
Test: demo mode using `-e ntn true` mobile command
Bug: 337103875
Flag: NONE
Change-Id: Id003d57662c682bab150795782827f600063243e
(cherry picked from commit 8d2b4221267ad70b30dcc15f7429c9b19eb84eb3)