Do not update TFParentInfo while invisible
WM Core sends the TFParentInfo updates to the client whenever
a Task is visible. That not just updates the visible tasks, but
also updates the TFParentInfo of the invisible ones. Though,
the TFContainer of the invisible Tasks are not updated.

The next time the Tasks become visible. The only change of the
TFParentInfo is the visibility, while the configuration stays
unchanged (due to it was already updated while the Task was
invisible). Therefore, the TFContainers are not updated using
the latest TFParentInfo.

Bug: 338153080
Test: atest SplitControllerTest
Change-Id: Ibb719519e2ab075cc44fb7273409947f1a8aac52
Merged-In: Ibb719519e2ab075cc44fb7273409947f1a8aac52
(cherry picked from commit 10a5886246dbd6dc214667ffc6c064b0a233fbd7)