Remove RemoteException handlings from AutofillSuggestionsController
This is a follow up CL to our previous CL [1], which introduced

  InlineSuggestionsRequestCallbackDecorator

that wraps

  InlineSuggestionsRequestCallbackImpl

to intercept IPC calls then inject special logic.

The problem is that when forwarding IPC calls within system_server
process, the AIDL-generated method stub forces us to handle
RemoteException, which actually never happen if the target Binder
object is in the same process. To clarify this nuance, this CL
introduces a plain Java interface

  com.android.internal.inputmethod.InlineSuggestionsRequestCallback

as a mirror of IInlineSuggestionsRequestCallback.

As you can see in this CL, doing so eliminates many RemoteException.

This is a mechanical code change with no observable behavior.

 [1]: I3d81b1d56989e46f65f9b9fc6e497952486335b5
      82961cf41ab39781a63dd1804c4d73e7b12eedd2

Bug: 339358344
Test: atest CtsAutoFillServiceTestCases
Test: atest CtsInputMethodTestCases
Change-Id: I25a75b47aaed29a42b153a7874e5bf26ca89df98