Move Trace Controlling to within Traceur app.
This is going to follow SELinux policy restrictions and allow the Record
Issue Quick Settings Tile to be pushed to production. It works now via a
bound service.

Also merging in change to use preset trace configurations, as this
limits the inter-app API required for this change.

Bug: 305049544
Test: Tested this locally.
Flag: com.android.systemui.record_issue_qs_tile STAGING

Change-Id: If5c3cdde9d139e11c68b44344ff0e86285b2b486