Add vm name and version to am capabilities
This CL is part of the go/kddm project, to deprecate DDM protocol.
HELO packet currently return the VM name and version, this CL allows
to retrieves the same invo from a tool accessible from shell.

The data is retrieved the same way it is retrieved in DdmHandleHello.
handleHELO (System.getproperty).

Test: NA
Bug: go/kddm

Change-Id: Id5cf9bf768f397bf3595e577f0da7a001b7ec394