Update NoMan's ShortcutHelper to use ShortcutChangeCallback
LauncherApps.Callback isn't user aware so we wouldn't hear about
changes to shortcuts on different user accounts. This was found out
due to bubble CTS not passing on HSUM.

ShortcutChangeCallback is user aware, so migrate to this. Also update
the keys used for shortcuts stored to be package|userId. Added some
tests around this.

Flag: EXEMPT bugfix
Test: atest ShortcutHelperTest NotificationManagerServiceTest CtsNotificationTestCases:NotificationManagerBubbleTest CtsNotificationTestCases:NotificationManagerTest
Test: ran the CTS tests on a secondary user account
Bug: 330507887
Change-Id: I610545fdcdff330d0ef625abd1a6ce0199e7bd08