Simplify IMMS.MyPackageMonitor
With my previous changes [1][2], long and/or frequent operations in

 MyPackageMonitor#onFinishPackageChangesInternal()

are expected to be no longer contributing to UI janks. Now it is time
to focus more on the correctness rather than trying to optimize things
in a complicated way. Complicated optimizations make it difficult for
us to diagnose issues like Bug 335281466.

This CL basically replaces my past optimizations [3][4] in a more
strict and straightforward way. We first create a new InputMethodMap
based on the latest results returned from the PackageManagerService,
then see if there is any actual data change from the current
InputMethodMap. If those two InputMethodMap instances are exactly the
same, it means that no further update is necessary.

There should be no semantic observable change.

 [1]: I263cd49dd4d64b64136acc3dad469f83a862ce97
      b04b6c584c991231827a68a85953423b7fed222c
 [2]: Icc84fe9d0d3b5842b4d5ee51b12b0d3cd10a1d24
      d8aed1d69d916dc7b23137d1f7f06229eaa37bf0
 [3]: I7b69c349318ce06a48d03a4468cf2c45bfb73dc2
      105bc1ceebd08e169713215dc1c7b8e536dad85b
 [4]: I063688297156188f68fe0b55a46d72f2e811dc88
      7091860c5dbb231a43b1623b8a077b2f40bbacbf

Bug: 329703038
Test: presubmit
Change-Id: I6e9de6ff5c6dd55e3c40f23586110cb713c62851