Status Bar - Fix privacy dot being in the wrong corner after fold/unfold
When unfolding and folding, the PrivacyDotViewController will be reused
and #initialize will be called with new views.

For the views to be in the correct position, their gravity has to be
set.
The fix is to make sure the gravity is updated when we have new views.

Fixes: 339335643
Test: Unit tests in this CL
Test: Manually
Flag: com.android.systemui.privacy_dot_unfold_wrong_corner_fix
Change-Id: Id37da91b04cf5fa956f3d6d7f254e9c2634a2563