Remove number code from StatusBarIconView
We had this to show an overlay with a number on notifications, but we
disabled that 13 years ago (ag/123288). This removes the dead code and
related resources that aren't used elsewhere.

I wanted to also remove ic_notification_overlay since we're not using
it, but it's technically a public resource and is used in some external
projects so that's probably not safe to do.

Similarly, two other resources are public so I just marked them as
deprecated to make it clear we don't want to use them anymore.

Bug: 315143361
Test: builds
Flag: EXEMPT removing dead code
Change-Id: I30e6c624b90c808e58f1c5c60b502517ef0011f7