Fix live timer activity launches
We add a new method to ActivityStarter to bypass the normal
showWhenLocked checks we do when starting an activity. This is due to
the fact that timers use a trampoline activity which is not
showWhenLocked, which then triggers a showWhenLocked activity. We
therefore do not want to trigger auth for these activities.

As this change is no longer a breaking change, the flag guarding this is
also removed.

Fixes: 345741071
Test: atest WidgetInteractionHandlerTest
Test: atest SmartspaceInteractionHandlerTest
Flag: com.android.systemui.communal_hub
Change-Id: I28dc5169642df8672578d28f9e3c47840277a32c