Position PiP animation based on actual coordinates
In the previous patchset, we assumed the left/top to be 0/0 which is not
always the case. This could break, for instance, in ActivityEmbedding
setup, where the pip-ing app is on the left side.

Changed also to use the Math.floor value for positions at the beginning
of the animation to avoid rounding errors (1 pixel off).

Flag: NONE test fix
Bug: 345327260
Test: atest FlickerTestsOther:SecondaryActivityEnterPipTest \
            WMShellFlickerTestsPip1 \
            WMShellFlickerTestsPip2 \
            WMShellFlickerTestsPip3 \
            PipAnimationControllerTest
Change-Id: I74739af591dd546c8cb35123fa580cfc6625fca4