Allow SHELL_UID to print FRP status when secrets are excluded
Bugreports taken by the `adb bugreport` command seem to
be executed by SHELL_UID (2000). Before this change, this meant
that the PersistentDataBlockService's dump would not include
the FRP status because SHELL_UID was not allowed.

Additionally, the OEM Unlock state and the Flash Lock state are
excluded when the caller is SHELL_UID because it seems to lack
the required permissions and throws an exception.

Bug: 349170452
Flag: EXEMPT bugfix
Test: Manually tested by running `adb shell dumpsys
persistent_data_block`
while not root
Ignore-AOSP-First: Depends on internal changes that can't go into AOSP
yet.

Change-Id: Icda8348c730ec869196839439ce1620d56a9339e