Reset stylus handwriting after IME crashed
Ideally, when handwriting is initiated, the IME should call back to
IMMS after finishing the handwriting, but when the IME crashes
abnormally, this callback will not be executed according to the
expected logic, which will result in the spy window not being removed
properly, and the user will no longer be able to operate the screen.

Here, we call scheduleResetStylusHandwriting after service disconnect
to solve this problem.


Bug: 351289050

Test: Start handwriting and keep stylus press on screen, execute `adb shell am crash com.google.android.inputmethod.latin.canary`, then swipe on screen, expect everything to be fine.

Change-Id: Ib19eac173b4455cda99c75544ca6f9c70e8a322f
Signed-off-by: Linnan Li <lilinnan@xiaomi.corp-partner.google.com>