Unify IMMS#onSecureSettingsChangedLocked() behaviors among users
This is a follow up CL to my previous CLs [1], which made our secure
settings observer multi-user aware.

What is remaining is an early exit case at the top of

  InputMethodManagerService#onSecureSettingsChangedLocked(),

which basically makes the above callback no-op when called for
background users when IMMS#mConcurrentMultiUserModeEnabled is false.

With this CL the above early exit rule is removed, meaning that
updates on secure settings will be propagated into downstream logic
even if called for background users.

 [1]: I35253d3964f37f3e13888c0017a86449cf1de8ac
      8da06000fd5e55d6d2aa252906e70d86e7e0fc45
 [2]: I27f94ca795f3035b5132a28fbbec669122d42b87
      3389daa8a025bbd10eae8169210a5a87d5760aec

Bug: 342027196
Bug: 350386877
Flag: android.view.inputmethod.concurrent_input_methods
Test: presubmit
Change-Id: I039816fa3e2acfe3cada3fb237c01560d474aee8