Remove ZeroJankProxy.Callback#isInputShownLocked()
This reworks our previous CL [1], which addressed an observable
unintentional behavior change when ZeroJankProxy is enabled, in a way
that ZeroJankProxy.Callback does not need to have

  ZeroJankProxy.Callback#isInputShownLocked(),

which will soon need to become concurrent multi-user aware.

The problem is when it becomes concurrent multi-user aware,
ZeroJankProxy needs to know whose IME is in use, which is tricky
because the IME user ID is dynamically resolved with

  InputMethodManagerService#resolveImeUserIdLocked().

The core logic should remain unchanged.

 [1]: Ie2f262128eb32db5a7922373fdfaebd2945ea36a
      d172f44ba0b77e09488782ec24f767e5969c8535

Bug: 349904272
Test: presubmit
Flag: android.view.inputmethod.use_zero_jank_proxy
Change-Id: Ib719f6018965c3554c8c2d8c2050ad51962d8a25