Fix race condition between service restart and unbind.
When a service crashed and all its clients unbind at the same time, if
the unbind happen to happen after the restart being scheduled, restart
will be cancelled. This is problematic for services that are started by
startService().

This change explicitly keep the service's restart status despite that
all connections are removed if the service is started by startService().

Flag: EXEMPTED bug fix
Bug: 293357096
Fix: 293357096
Test: atest ServiceTest
Test: manual test on watch. Kill AmbientService and observed that it
      always restarts.
Change-Id: Ie1ab8b8906009d3f39d7f04b267936a77a0f3f89