Don't return generic policy definitions
Methods modified in this CL are used exclusively to build policies that
are settable by admins. A generic policy cannot be set - it is missing
an argument (e.g. a package name), and can't really be serialized
because Xml.attribute will throw when given a null value. Thus when
these methods are used with null arguments, it is a programming error.

Bug: 335663055
Test: TH
Flag: EXEMPT trivial change
Change-Id: Idd7f16bf215c35d1e860612c8d901711bf3b476d