Fix handwriting trigger on wrong textfield
In order to start handwriting faster, we have introduced the ability to
start handwriting based on the EditText that currently has the focus,
if the focus is switched to an EditText that doesn't support
handwriting, we should clear the current focus View, otherwise there
will be a difference between the EditText that has the focus and the
actual EditText will be inconsistent with the actual EditText. This
will cause handwriting to be triggered incorrectly for some EditTexts.

Bug: 353651951
Test: atest StylusHandwritingTest
Flag: android.view.inputmethod.initiation_without_input_connection

Signed-off-by: Linnan Li <lilinnan@xiaomi.corp-partner.google.com>
Merged-In: Id5256724ed0d454bad3e81ec3f28d009ec0ae252
Change-Id: Id5256724ed0d454bad3e81ec3f28d009ec0ae252