Remove check for frame size
Previously there was a check to ensure that all frames have the exact same size and share the same hotspot.
A crash occurred if this condition was not satisfied,
which was causing presubmit to fail due to test failures.
Removing this crash condition to fix presubmit and added a TODO for a
more permanent solution.

Bug: 361232935
Test: Presubmit
Flag: EXEMPT fix crash
Change-Id: I256b1f470442395876bd6fd273d87be72578d276