Remove explicit creation of SurfaceSession in WMS
Because it is a stateless connection object to SF, it is unnecessary
to have individual connections from DisplayContent and the Session
per-process.

There were already several places use null for SurfaceControl.Builder,
so the singleton instance SurfaceComposerClient::getDefault() is
already in use. Then just use that one. That reduces unnecessary
IPC to SF when creating surface for new processes.

Bug: 159103089
Flag: EXEMPT reduce unnecessary invocation
Test: atest WindowManagerServiceTests

Change-Id: Iab7e708d064dd42d39d3ab8cf3e08d657b926c68