Revert "Move showing keyguard after the UserSwitchObservers."
This reverts commit 8038b18cc4f624a6fc70924b2a39159bc5fca0a4.

Reason for revert: This CL caused b/360838273, we need revert it and show the keyguard at the beginning of the user switch as we were doing before. Otherwise, it can cause security issues such as the ones described in b/360838273#comment22.

This needs to be submitted together with the proper fix in `WindowManagerService.lockDeviceNow()`.

Bug: 331853529
Fixes: 360838273

Change-Id: I8f25255ac204160092fb88c9f099c697b222dc88