Modify test tag on ModeTile to include state
We clear the text & content description on the semantics node for the mode tile's subtext in order to reduce duplicate state information for screen readers; but this caused the OpenModesDialog test to fail. This change modifies the test tag so that it reflects the state, allowing UiAutomator tests to observe the state of the tile.

Bug: 363341778
Bug: 359845144
Test: OpenModesDialog
Test: ModesDialogViewModelTest
Test: manual with talkback
Flag: android.app.modes_ui
Change-Id: I7b26ec36d5302f3de8efb3b3bb4f4fe3490174f7