Add multi-display support to focus switching in shell transitions
With [1], a transition is created when a task is moved to front
even without any visibility change. This change extends this to
multi-display scenarios.

Task focus is managed per display, which means that, when a top
task on a non-top display is clicked, there's no task focus switch,
(but only display focus switch). The current transition system
doesn't track dipslay focus, so in such cases no transition
happens, and there's no way for Shell to know when display focus
is switched.

With this change, top display is tracked in the same way as top
tasks, so a TO_TOP transition is delivered to Shell even when
there's no task switch inside a display is involved.

[1] If21d076eed4db88139ffc8a7c4c018c2ef5aad93

Bug: 274696524
Test: TransitionTests
Flag: com.android.window.flags.enable_display_focus_in_shell_transitions
Change-Id: I3f74a169eaea5d71a69132a4aa5693fd27bc372c