Move a11ychecker sources to an isolated dependency
Per the discussions on ag/28470431 and ag/28136135, a11ychecker's "aatf" dependency can't be included in the system server since it pulls in test-only dependency. The goal of this CL is to remove the dependency from the system server while this is resolved.

Bug: 364326163
Flag: com.android.server.accessibility.enable_a11y_checker_logging
Test: N/A
Change-Id: I0cf2cf6ddce2eff0a70d36d90c9b79101db9dab6