[SB] Update view model to hide status bar when in keyguard or shade.
Instead of trying to mimic all the incredibly complicated visibility
logic from CollapsedStatusBarFragment into CollapsedStatusBarViewModel,
this CL just starts with the basic facts: Hide the status bar when
lockscreen, shade, or bouncer is visible; show the status bar if the
lockscreen is occluded or none of those surfaces are visible.

We'll add back in the edge cases from CSBF as-needed, but this starts
with just the basics.

Bug: 364360986
Flag: com.android.systemui.status_bar_simple_fragment
Test: Verify status bar appears on home screen and in apps but not when
on lockscreen, bouncer, or shade
Test: Launch occluding activity that requires status bar over lockscreen
(e.g. with lockscreen shortcut) -> verify home status bar appears
Test: With status_bar_simple_fragment flag off, verify everything still
works as before and there's no crashes
Test: atest CollapsedStatusBarViewModelTest
CollapsedStatusBarFragmentTest

Change-Id: I053e6bc2e560e060dd2d5cf1e276ba97d0443693